(refactor): Break up big code files and adjust rate-limiting
