+++ .env.example
@@ -1,26 +1,42 @@
-# Database
+# ─── Database ───────────────────────────────────────────────────────
+# PostgreSQL credentials used by docker-compose and the backend.
 DB_USER=postgres
-DB_PASSWORD=your_secure_password_here
+DB_PASSWORD=your_secure_password_here          # REQUIRED — use a strong random password (32+ chars)
 
-# MinIO (Object Storage)
+# ─── MinIO (S3-compatible Object Storage) ───────────────────────────
+# Credentials for the MinIO container that stores uploaded media.
 MINIO_USER=minio_user
-MINIO_PASSWORD=your_secure_password_here
+MINIO_PASSWORD=your_secure_password_here       # REQUIRED — min 8 chars
 
-# S3 Storage
+# ─── S3 Storage ─────────────────────────────────────────────────────
+# AWS region for the S3/MinIO bucket (used by @payloadcms/storage-s3).
 S3_REGION=eu-central-1
 
-# Backend
-PAYLOAD_SECRET=your_jwt_secret_here
-DOMAIN_NAME=localhost
-DOMAIN_NAME_WWW=www.localhost
-IP_ADDRESS=http://localhost
-ALTCHA_SECRET=your_altcha_secret_key_here
+# ─── Backend / Payload CMS ─────────────────────────────────────────
+# JWT signing secret for Payload admin sessions.
+PAYLOAD_SECRET=your_jwt_secret_here            # REQUIRED — use a random 32+ char string (e.g. openssl rand -hex 32)
 
-# Email (SMTP)
-SMTP_HOST=smtp.example.com
-SMTP_PORT=587
-SMTP_SECURE=false
-SMTP_PASS=your_smtp_password
-CONTACT_EMAIL=info@shi-sei.nl
-TRIAL_LESSON_EMAIL=proefles@shi-sei.nl
-BCC_EMAIL=shi-sei@mail.com
+# Domain names used to build CORS origins, CSRF protection, and canonical URLs.
+DOMAIN_NAME=localhost                          # Primary domain (without protocol)
+DOMAIN_NAME_WWW=www.localhost                  # WWW variant
+IP_ADDRESS=http://localhost                    # Full base URL (with protocol) used as PAYLOAD_PUBLIC_SERVER_URL
+
+# Secret key for Altcha CAPTCHA challenge generation and verification.
+ALTCHA_SECRET=your_altcha_secret_key_here      # REQUIRED — generate with: openssl rand -hex 32
+
+# ─── Email (SMTP) ──────────────────────────────────────────────────
+# Both CONTACT_EMAIL and TRIAL_LESSON_EMAIL share the same SMTP host/port/pass
+# but each authenticates with its own address as SMTP_USER.
+SMTP_HOST=smtp.example.com                     # REQUIRED — SMTP server hostname
+SMTP_PORT=587                                  # SMTP port (587 for STARTTLS, 465 for SSL)
+SMTP_SECURE=false                              # "true" for port 465 (implicit TLS), "false" for STARTTLS on 587
+SMTP_PASS=your_smtp_password                   # REQUIRED — SMTP password (shared by both email accounts)
+
+# Email address for contact and enrollment forms (sends + receives).
+CONTACT_EMAIL=info@shi-sei.nl                  # REQUIRED
+
+# Email address for trial lesson requests (sends + receives).
+TRIAL_LESSON_EMAIL=proefles@shi-sei.nl         # REQUIRED
+
+# Optional BCC recipient for club-facing emails (not user confirmations).
+BCC_EMAIL=shi-sei@mail.com                     # OPTIONAL — leave empty to disable
