+++ frontend/src/components/EnrollmentForm.tsx
@@ -2,79 +2,23 @@ import { useState, useEffect, useRef, useCallback } from 'react';
 import { Link } from 'react-router-dom';
 import { useLanguage } from '../contexts/LanguageContext';
 import { api } from '../lib/api';
-import { Check, AlertCircle, Loader, ArrowRight, X } from 'lucide-react';
+import { Check, AlertCircle, Loader, ArrowRight } from 'lucide-react';
 import { FillButton } from './FillButton';
 import { CustomSelect } from './CustomSelect';
-import { IbanInput } from './IbanInput';
-import { PhoneInput } from './PhoneInput';
 import { SignaturePad } from './SignaturePad';
 import { isValidEmail, isValidPhone, isValidIban, isValidPostalCode } from '../lib/validation';
 import 'altcha';
-import { useFocusTrap } from '../hooks/useFocusTrap';
 import { useCountdown } from '../hooks/useCountdown';
-
-interface EnrollmentFormData {
-  firstName: string;
-  middleName: string;
-  lastName: string;
-  email: string;
-  phone: string;
-  dateOfBirth: string;
-  address: {
-    street: string;
-    houseNumber: string;
-    postalCode: string;
-    city: string;
-  };
-  guardianName: string;
-  experience: string;
-  judoGrade: string;
-  medicalInfo: string;
-  preferredTrainingDays: string[];
-  remarks: string;
-  paymentMethod: 'regular' | 'ooievaarspas';
-  ooievaarspasNumber: string;
-  bankAccount?: {
-    accountHolder: string;
-    iban: string;
-  };
-}
-
-const formatPostalCode = (value: string): string => {
-  const stripped = value.replace(/\s/g, '').toUpperCase().slice(0, 6);
-  return stripped.length > 4 ? stripped.slice(0, 4) + ' ' + stripped.slice(4) : stripped;
-};
-
+import { PersonalInfoSection } from './enrollment/PersonalInfoSection';
+import { AddressSection } from './enrollment/AddressSection';
+import { PaymentSection } from './enrollment/PaymentSection';
+import { ConfirmationModal } from './enrollment/ConfirmationModal';
+import { INITIAL_FORM_DATA, formatPostalCode } from './enrollment/types';
+import type { EnrollmentFormData } from './enrollment/types';
 
 export const EnrollmentForm = () => {
   const { t } = useLanguage();
-  const [formData, setFormData] = useState<EnrollmentFormData>({
-    firstName: '',
-    middleName: '',
-    lastName: '',
-    email: '',
-    phone: '',
-    dateOfBirth: '',
-    address: {
-      street: '',
-      houseNumber: '',
-      postalCode: '',
-      city: '',
-    },
-    guardianName: '',
-    experience: 'beginner',
-    judoGrade: '',
-    medicalInfo: '',
-    preferredTrainingDays: [],
-    remarks: '',
-    paymentMethod: 'regular',
-    ooievaarspasNumber: '',
-    bankAccount: {
-      accountHolder: '',
-      iban: '',
-    },
-  });
-
+  const [formData, setFormData] = useState<EnrollmentFormData>({ ...INITIAL_FORM_DATA });
   const [submitting, setSubmitting] = useState(false);
   const [submitted, setSubmitted] = useState(false);
   const [showConfirmation, setShowConfirmation] = useState(false);
@@ -83,21 +27,13 @@ export const EnrollmentForm = () => {
   const [agreedToTerms, setAgreedToTerms] = useState(false);
   const [signature, setSignature] = useState<string | null>(null);
   const altchaRef = useRef<HTMLElement>(null);
-  const confirmModalRef = useRef<HTMLDivElement>(null);
   const triggerElementRef = useRef<Element | null>(null);
   const successRef = useRef<HTMLDivElement>(null);
-  useFocusTrap(confirmModalRef, showConfirmation, () => setShowConfirmation(false));
   const [touched, setTouched] = useState<Record<string, boolean>>({});
 
   const handleBlur = (e: React.FocusEvent<HTMLInputElement | HTMLTextAreaElement | HTMLSelectElement>) => {
     setTouched((prev) => ({ ...prev, [e.target.name]: true }));
   };
-
-  const emailError = touched.email && formData.email.trim() && !isValidEmail(formData.email) ? t('common.invalidEmail') : null;
-  const phoneNumber = formData.phone.replace(/^\+\d{1,4}-/, '').trim();
-  const phoneError = touched.phone && phoneNumber && !isValidPhone(formData.phone) ? t('common.invalidPhone') : null;
-  const postalCodeError = touched['address.postalCode'] && formData.address.postalCode.trim() && !isValidPostalCode(formData.address.postalCode) ? t('common.invalidPostalCode') : null;
-  const ibanError = touched['bankAccount.iban'] && formData.bankAccount?.iban?.trim() && !isValidIban(formData.bankAccount.iban) ? t('common.invalidIban') : null;
 
   const handleSignatureChange = useCallback((dataUrl: string | null) => {
     setSignature(dataUrl);
@@ -150,32 +86,7 @@ export const EnrollmentForm = () => {
 
       if (response.data.success) {
         setSubmitted(true);
-        setFormData({
-          firstName: '',
-          middleName: '',
-          lastName: '',
-          email: '',
-          phone: '',
-          dateOfBirth: '',
-          address: {
-            street: '',
-            houseNumber: '',
-            postalCode: '',
-            city: '',
-          },
-          guardianName: '',
-          experience: 'beginner',
-          judoGrade: '',
-          medicalInfo: '',
-          preferredTrainingDays: [],
-          remarks: '',
-          paymentMethod: 'regular',
-          ooievaarspasNumber: '',
-          bankAccount: {
-            accountHolder: '',
-            iban: '',
-          },
-        });
+        setFormData({ ...INITIAL_FORM_DATA });
         setAltchaPayload(null);
         setSignature(null);
         if (altchaRef.current) {
@@ -205,39 +116,42 @@ export const EnrollmentForm = () => {
 
     if (name.startsWith('address.')) {
       const addressField = name.split('.')[1];
-      setFormData({
-        ...formData,
-        address: { ...formData.address, [addressField]: value },
-      });
+      setFormData((prev) => ({
+        ...prev,
+        address: { ...prev.address, [addressField]: value },
+      }));
     } else if (name.startsWith('bankAccount.')) {
       const field = name.split('.')[1];
-      setFormData({
-        ...formData,
+      setFormData((prev) => ({
+        ...prev,
         bankAccount: {
-          accountHolder: formData.bankAccount?.accountHolder || '',
-          iban: formData.bankAccount?.iban || '',
+          accountHolder: prev.bankAccount?.accountHolder || '',
+          iban: prev.bankAccount?.iban || '',
           [field]: value
         },
-      });
+      }));
     } else {
-      const updated = { ...formData, [name]: value };
-      if (name === 'dateOfBirth' && value) {
-        const dob = new Date(value);
-        const cutoff = new Date();
-        cutoff.setFullYear(cutoff.getFullYear() - 18);
-        if (dob <= cutoff) updated.guardianName = '';
-      }
-      setFormData(updated);
+      setFormData((prev) => {
+        const updated = { ...prev, [name]: value };
+        if (name === 'dateOfBirth' && value) {
+          const dob = new Date(value);
+          const cutoff = new Date();
+          cutoff.setFullYear(cutoff.getFullYear() - 18);
+          if (dob <= cutoff) updated.guardianName = '';
+        }
+        return updated;
+      });
     }
   };
 
-  const isAdult = (() => {
-    if (!formData.dateOfBirth) return false;
-    const dob = new Date(formData.dateOfBirth);
-    const cutoff = new Date();
-    cutoff.setFullYear(cutoff.getFullYear() - 18);
-    return dob <= cutoff;
-  })();
+  const handleDaysChange = (day: string) => {
+    setFormData((prev) => ({
+      ...prev,
+      preferredTrainingDays: prev.preferredTrainingDays.includes(day)
+        ? prev.preferredTrainingDays.filter((d) => d !== day)
+        : [...prev.preferredTrainingDays, day],
+    }));
+  };
 
   const isFormValid =
     formData.firstName.trim() !== '' &&
@@ -261,21 +175,6 @@ export const EnrollmentForm = () => {
     agreedToTerms &&
     signature !== null &&
     altchaPayload !== null;
-
-  const handleDaysChange = (day: string) => {
-    const currentDays = formData.preferredTrainingDays;
-    if (currentDays.includes(day)) {
-      setFormData({
-        ...formData,
-        preferredTrainingDays: currentDays.filter((d) => d !== day),
-      });
-    } else {
-      setFormData({
-        ...formData,
-        preferredTrainingDays: [...currentDays, day],
-      });
-    }
-  };
 
   const { remaining, progress } = useCountdown(submitted, 20);
 
@@ -301,6 +200,8 @@ export const EnrollmentForm = () => {
     );
   }
 
+  const sectionProps = { formData, handleChange, handleBlur, t };
+
   return (
     <form onSubmit={handleSubmit} className="space-y-8">
       {error && (
@@ -311,182 +212,10 @@ export const EnrollmentForm = () => {
       )}
 
       {/* Personal Information */}
-      <div>
-        <h3 className="text-lg font-bold text-judo-dark mb-4">{t('enrollment.form.personalInfo')}</h3>
-        <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-          <div>
-            <label className="block text-sm font-medium text-judo-dark mb-2">
-              {t('enrollment.form.firstName')} *
-            </label>
-            <input
-              type="text"
-              name="firstName"
-              value={formData.firstName}
-              onChange={handleChange}
-              required
-              placeholder={t('placeholder.firstName')}
-              className="w-full px-4 py-2 text-sm sm:text-base border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-judo-red focus:border-transparent"
-            />
-          </div>
-          <div>
-            <label className="block text-sm font-medium text-judo-dark mb-2">
-              {t('enrollment.form.middleName')}
-            </label>
-            <input
-              type="text"
-              name="middleName"
-              value={formData.middleName}
-              onChange={handleChange}
-              placeholder={t('placeholder.middleName')}
-              className="w-full px-4 py-2 text-sm sm:text-base border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-judo-red focus:border-transparent"
-            />
-          </div>
-          <div className="md:col-span-2">
-            <label className="block text-sm font-medium text-judo-dark mb-2">
-              {t('enrollment.form.lastName')} *
-            </label>
-            <input
-              type="text"
-              name="lastName"
-              value={formData.lastName}
-              onChange={handleChange}
-              required
-              placeholder={t('placeholder.lastName')}
-              className="w-full px-4 py-2 text-sm sm:text-base border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-judo-red focus:border-transparent"
-            />
-          </div>
-          <div>
-            <label className="block text-sm font-medium text-judo-dark mb-2">
-              {t('enrollment.form.dateOfBirth')} *
-            </label>
-            <input
-              type="date"
-              name="dateOfBirth"
-              value={formData.dateOfBirth}
-              onChange={handleChange}
-              max={(() => { const d = new Date(); d.setFullYear(d.getFullYear() - 4); return d.toISOString().split('T')[0]; })()}
-              required
-              className="w-full px-4 py-2 text-sm sm:text-base border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-judo-red focus:border-transparent"
-            />
-          </div>
-          <div>
-            <label className="block text-sm font-medium text-judo-dark mb-2">
-              {t('enrollment.form.guardianName')}
-            </label>
-            <input
-              type="text"
-              name="guardianName"
-              value={formData.guardianName}
-              onChange={handleChange}
-              disabled={isAdult}
-              placeholder={t('enrollment.form.guardianNamePlaceholder')}
-              className="w-full px-4 py-2 text-sm sm:text-base border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-judo-red focus:border-transparent disabled:bg-gray-100 disabled:text-gray-400 disabled:cursor-not-allowed"
-            />
-          </div>
-          <div>
-            <label className="block text-sm font-medium text-judo-dark mb-2">
-              {t('enrollment.form.email')} *
-            </label>
-            <input
-              type="email"
-              name="email"
-              id="enrollment-email"
-              value={formData.email}
-              onChange={handleChange}
-              onBlur={handleBlur}
-              required
-              aria-invalid={emailError ? true : undefined}
-              aria-describedby={emailError ? 'enrollment-email-error' : undefined}
-              placeholder={t('placeholder.email')}
-              className={`w-full px-4 py-2 text-sm sm:text-base border rounded-lg focus:outline-none focus:ring-2 focus:ring-judo-red focus:border-transparent ${emailError ? 'border-red-400' : 'border-gray-300'}`}
-            />
-            {emailError && <p id="enrollment-email-error" role="alert" className="text-sm text-red-600 mt-1">{emailError}</p>}
-          </div>
-          <div>
-            <label className="block text-sm font-medium text-judo-dark mb-2">
-              {t('enrollment.form.phone')} *
-            </label>
-            <PhoneInput
-              id="enrollment-phone"
-              value={formData.phone}
-              onChange={(val) => setFormData({ ...formData, phone: val })}
-              onBlur={() => setTouched((prev) => ({ ...prev, phone: true }))}
-              required
-              hasError={!!phoneError}
-              aria-describedby={phoneError ? 'enrollment-phone-error' : undefined}
-            />
-            {phoneError && <p id="enrollment-phone-error" role="alert" className="text-sm text-red-600 mt-1">{phoneError}</p>}
-          </div>
-        </div>
-      </div>
+      <PersonalInfoSection {...sectionProps} touched={touched} setTouched={setTouched} setFormData={setFormData} />
 
       {/* Address */}
-      <div>
-        <h3 className="text-lg font-bold text-judo-dark mb-4">{t('enrollment.form.address')}</h3>
-        <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-          <div className="md:col-span-2">
-            <label className="block text-sm font-medium text-judo-dark mb-2">
-              {t('enrollment.form.street')} *
-            </label>
-            <input
-              type="text"
-              name="address.street"
-              value={formData.address.street}
-              onChange={handleChange}
-              required
-              placeholder={t('placeholder.street')}
-              className="w-full px-4 py-2 text-sm sm:text-base border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-judo-red focus:border-transparent"
-            />
-          </div>
-          <div>
-            <label className="block text-sm font-medium text-judo-dark mb-2">
-              {t('enrollment.form.houseNumber')} *
-            </label>
-            <input
-              type="text"
-              name="address.houseNumber"
-              value={formData.address.houseNumber}
-              onChange={handleChange}
-              required
-              placeholder={t('placeholder.houseNumber')}
-              className="w-full px-4 py-2 text-sm sm:text-base border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-judo-red focus:border-transparent"
-            />
-          </div>
-          <div>
-            <label className="block text-sm font-medium text-judo-dark mb-2">
-              {t('enrollment.form.postalCode')} *
-            </label>
-            <input
-              type="text"
-              name="address.postalCode"
-              id="enrollment-postalcode"
-              value={formData.address.postalCode}
-              onChange={handleChange}
-              onBlur={handleBlur}
-              required
-              aria-invalid={postalCodeError ? true : undefined}
-              aria-describedby={postalCodeError ? 'enrollment-postalcode-error' : undefined}
-              placeholder="1234 AB"
-              className={`w-full px-4 py-2 text-sm sm:text-base border rounded-lg focus:outline-none focus:ring-2 focus:ring-judo-red focus:border-transparent ${postalCodeError ? 'border-red-400' : 'border-gray-300'}`}
-            />
-            {postalCodeError && <p id="enrollment-postalcode-error" role="alert" className="text-sm text-red-600 mt-1">{postalCodeError}</p>}
-          </div>
-          <div className="md:col-span-2">
-            <label className="block text-sm font-medium text-judo-dark mb-2">
-              {t('enrollment.form.city')} *
-            </label>
-            <input
-              type="text"
-              name="address.city"
-              value={formData.address.city}
-              onChange={handleChange}
-              required
-              placeholder={t('placeholder.city')}
-              className="w-full px-4 py-2 text-sm sm:text-base border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-judo-red focus:border-transparent"
-            />
-          </div>
-        </div>
-      </div>
+      <AddressSection {...sectionProps} touched={touched} />
 
       {/* Judo Experience */}
       <div>
@@ -585,92 +314,7 @@ export const EnrollmentForm = () => {
       </div>
 
       {/* Payment Method */}
-      <div className="payment-section bg-yellow-50 border-2 border-yellow-200 rounded-lg p-4 sm:p-6">
-        <h3 className="text-lg font-bold text-judo-dark mb-4">{t('enrollment.form.paymentMethod')}</h3>
-        <p className="text-sm text-gray-600 mb-4">{t('enrollment.form.paymentMethodNote')}</p>
-
-        {/* Ooievaarspas Checkbox */}
-        <label className="flex items-start gap-3 p-4 -mx-4 sm:mx-0 sm:border-2 border-yellow-300 sm:rounded-lg cursor-pointer hover:bg-yellow-100 transition-colors mb-4">
-          <input
-            type="checkbox"
-            checked={formData.paymentMethod === 'ooievaarspas'}
-            onChange={(e) => {
-              setFormData({
-                ...formData,
-                paymentMethod: e.target.checked ? 'ooievaarspas' : 'regular',
-              });
-            }}
-            className="w-5 h-5 text-judo-red focus:ring-judo-red mt-1"
-          />
-          <div>
-            <span className="font-bold text-judo-dark block">{t('enrollment.form.ooievaarspas')}</span>
-            <span className="text-sm text-gray-600">{t('enrollment.form.ooievaarspasNote')}</span>
-          </div>
-        </label>
-
-        {/* Ooievaarspas Number - Only show when ooievaarspas is selected */}
-        {formData.paymentMethod === 'ooievaarspas' && (
-          <div className="bg-white -mx-4 sm:mx-0 p-4 sm:p-6 sm:border border-yellow-300 sm:rounded-lg mb-4">
-            <div>
-              <label className="block text-sm font-medium text-judo-dark mb-2">
-                {t('enrollment.form.ooievaarspasNumber')} *
-              </label>
-              <input
-                type="text"
-                name="ooievaarspasNumber"
-                value={formData.ooievaarspasNumber}
-                onChange={handleChange}
-                required
-                placeholder={t('enrollment.form.ooievaarspasNumberPlaceholder')}
-                className="w-full px-4 py-2 text-sm sm:text-base border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-judo-red focus:border-transparent"
-              />
-            </div>
-            <p className="text-sm text-yellow-700 mt-3">{t('enrollment.form.ooievaarspasCheckLocation')}</p>
-          </div>
-        )}
-
-        {/* Bank Account (Machtiging) - Only show if NOT using Ooievaarspas */}
-        {formData.paymentMethod === 'regular' && (
-          <div className="bg-white -mx-4 sm:mx-0 p-4 sm:p-6 sm:border border-gray-200 sm:rounded-lg">
-            <h4 className="font-bold text-judo-dark mb-3">{t('enrollment.form.machtiging')}</h4>
-            <p className="text-sm text-gray-600 mb-4">{t('enrollment.form.machtigingNote')}</p>
-            <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-              <div className="md:col-span-2">
-                <label className="block text-sm font-medium text-judo-dark mb-2">
-                  {t('enrollment.form.accountHolder')} *
-                </label>
-                <input
-                  type="text"
-                  name="bankAccount.accountHolder"
-                  value={formData.bankAccount?.accountHolder || ''}
-                  onChange={handleChange}
-                  required={formData.paymentMethod === 'regular'}
-                  placeholder={t('placeholder.accountHolder')}
-                  className="w-full px-4 py-2 text-sm sm:text-base border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-judo-red focus:border-transparent"
-                />
-              </div>
-              <div className="md:col-span-2">
-                <label className="block text-sm font-medium text-judo-dark mb-2">
-                  {t('enrollment.form.iban')} *
-                </label>
-                <IbanInput
-                  id="enrollment-iban"
-                  value={formData.bankAccount?.iban || ''}
-                  onChange={(val) => setFormData({
-                    ...formData,
-                    bankAccount: { accountHolder: formData.bankAccount?.accountHolder || '', iban: val },
-                  })}
-                  onBlur={() => setTouched((prev: Record<string, boolean>) => ({ ...prev, 'bankAccount.iban': true }))}
-                  required={formData.paymentMethod === 'regular'}
-                  hasError={!!ibanError}
-                  aria-describedby={ibanError ? 'enrollment-iban-error' : undefined}
-                />
-                {ibanError && <p id="enrollment-iban-error" role="alert" className="text-sm text-red-600 mt-1">{ibanError}</p>}
-              </div>
-            </div>
-          </div>
-        )}
-      </div>
+      <PaymentSection {...sectionProps} touched={touched} setTouched={setTouched} setFormData={setFormData} />
 
       {/* Signature */}
       <div>
@@ -741,92 +385,12 @@ export const EnrollmentForm = () => {
 
       {/* Confirmation Modal */}
       {showConfirmation && (
-        <div className="fixed inset-0 bg-black/50 z-50 flex items-center justify-center p-4" onClick={() => setShowConfirmation(false)}>
-          <div
-            ref={confirmModalRef}
-            role="dialog"
-            aria-modal="true"
-            aria-labelledby="enrollment-confirm-title"
-            className="bg-white rounded-2xl shadow-xl max-w-lg w-full max-h-[90vh] overflow-y-auto"
-            onClick={(e) => e.stopPropagation()}
-          >
-            <div className="flex items-center justify-between p-6 border-b border-gray-200">
-              <h3 id="enrollment-confirm-title" className="text-lg font-bold text-judo-dark">{t('enrollment.confirm.title')}</h3>
-              <button onClick={() => setShowConfirmation(false)} className="text-gray-400 hover:text-gray-600 p-1">
-                <X className="w-5 h-5" />
-              </button>
-            </div>
-
-            <div className="p-6 space-y-4">
-              <p className="text-sm text-gray-600">{t('enrollment.confirm.description')}</p>
-
-              <div className="bg-gray-50 rounded-lg p-4 space-y-2 text-sm">
-                <div className="flex justify-between">
-                  <span className="text-gray-500">{t('enrollment.form.firstName')}</span>
-                  <span className="font-medium text-judo-dark">{formData.firstName}</span>
-                </div>
-                {formData.middleName && (
-                  <div className="flex justify-between">
-                    <span className="text-gray-500">{t('enrollment.form.middleName')}</span>
-                    <span className="font-medium text-judo-dark">{formData.middleName}</span>
-                  </div>
-                )}
-                <div className="flex justify-between">
-                  <span className="text-gray-500">{t('enrollment.form.lastName')}</span>
-                  <span className="font-medium text-judo-dark">{formData.lastName}</span>
-                </div>
-                {formData.guardianName && (
-                  <div className="flex justify-between">
-                    <span className="text-gray-500">{t('enrollment.form.guardianName')}</span>
-                    <span className="font-medium text-judo-dark">{formData.guardianName}</span>
-                  </div>
-                )}
-                <div className="flex justify-between">
-                  <span className="text-gray-500">{t('enrollment.form.dateOfBirth')}</span>
-                  <span className="font-medium text-judo-dark">{formData.dateOfBirth}</span>
-                </div>
-                <div className="flex justify-between">
-                  <span className="text-gray-500">{t('enrollment.form.email')}</span>
-                  <span className="font-medium text-judo-dark">{formData.email}</span>
-                </div>
-                <div className="flex justify-between">
-                  <span className="text-gray-500">{t('enrollment.form.phone')}</span>
-                  <span className="font-medium text-judo-dark">{formData.phone}</span>
-                </div>
-                <div className="flex justify-between">
-                  <span className="text-gray-500">{t('enrollment.form.address')}</span>
-                  <span className="font-medium text-judo-dark text-right">
-                    {formData.address.street} {formData.address.houseNumber}<br />
-                    {formData.address.postalCode} {formData.address.city}
-                  </span>
-                </div>
-                <div className="flex justify-between">
-                  <span className="text-gray-500">{t('enrollment.form.paymentMethod')}</span>
-                  <span className="font-medium text-judo-dark">
-                    {formData.paymentMethod === 'ooievaarspas' ? t('enrollment.form.ooievaarspas') : t('enrollment.form.machtiging')}
-                  </span>
-                </div>
-              </div>
-            </div>
-
-            <div className="flex gap-3 p-6 border-t border-gray-200">
-              <button
-                type="button"
-                onClick={() => setShowConfirmation(false)}
-                className="flex-1 px-4 py-3 border border-gray-300 rounded-lg font-medium text-gray-700 hover:bg-gray-50 transition-colors"
-              >
-                {t('enrollment.confirm.cancel')}
-              </button>
-              <button
-                type="button"
-                onClick={handleConfirmSubmit}
-                className="flex-1 px-4 py-3 bg-judo-red text-white rounded-lg font-bold hover:bg-red-700 transition-colors"
-              >
-                {t('enrollment.confirm.submit')}
-              </button>
-            </div>
-          </div>
-        </div>
+        <ConfirmationModal
+          formData={formData}
+          t={t}
+          onConfirm={handleConfirmSubmit}
+          onCancel={() => setShowConfirmation(false)}
+        />
       )}
     </form>
   );
