+++ frontend/src/contexts/LanguageContext.tsx
@@ -1,4 +1,6 @@
 import { createContext, useContext, useState, useEffect, type ReactNode } from 'react';
+import translationsNl from '../i18n/nl';
+import translationsEn from '../i18n/en';
 
 type Language = 'nl' | 'en';
 
@@ -53,1188 +55,4 @@ export const LanguageProvider = ({ children }: LanguageProviderProps) => {
       {children}
     </LanguageContext.Provider>
   );
-};
-
-// Translation objects
-const translationsNl: Record<string, string> = {
-  // Navigation
-  'nav.home': 'Home',
-  'nav.about': 'Over Ons',
-  'nav.team': 'Het Team',
-  'nav.history': 'Historie',
-  'nav.locations': 'Locaties',
-  'nav.rules': 'VCP en Regels',
-  'nav.info': 'Informatie',
-  'nav.schedule': 'Rooster',
-  'nav.pricing': 'Tarieven',
-  'nav.exam': 'Examen Eisen',
-  'nav.enrollment': 'Inschrijven',
-  'nav.agenda': 'Agenda',
-  'nav.news': 'Nieuws',
-  'nav.media': 'Media',
-  'nav.contact': 'Contact',
-  'nav.trial': 'Gratis Proefles',
-
-  
-  // Hero
-  'hero.title': 'Judo',
-  'hero.subtitle': 'SHI-SEI SPORT',
-  'hero.description': 'De oudste judo club van Den Haag',
-  'hero.text': 'Wij geven judo les op maandag, woensdag, donderdag en zaterdag. Ervaar de kracht van discipline, respect en zelfvertrouwen in onze dojo.',
-  'hero.button': 'Probeer een Proefles',
-  'hero.scheduleButton': 'Bekijk het Rooster',
-  
-  // News
-  'news.latest': 'Laatste Nieuws',
-  'news.search': 'Zoek nieuws...',
-  'news.all': 'Alle nieuws',
-  'news.readMore': 'Lees meer',
-  'news.clickToRead': 'Klik om het volledige bericht te lezen...',
-  'news.loading': 'Laden...',
-  'news.notFound': 'Nieuws niet gevonden',
-  'news.back': 'Terug naar nieuws',
-  'news.backAll': 'Terug naar alle nieuws',
-  'news.noNews': 'Er is nog geen nieuws beschikbaar.',
-  'news.title': 'Laatste Nieuws',
-  'news.description': 'Blijf op de hoogte van het laatste nieuws en evenementen bij Shi-Sei Sport.',
-  'news.share': 'Deel dit artikel',
-  'news.shareCopied': 'Link gekopieerd!',
-  'news.noContent': 'Dit artikel heeft geen inhoud.',
-
-  // Schedule
-  'schedule.title': 'Lessen & Schema',
-  'schedule.description': 'Wij geven judo les op maandag, woensdag, donderdag en zaterdag',
-  'schedule.loading': 'Rooster laden...',
-  'schedule.error': 'Kon het rooster niet laden. Probeer het later opnieuw.',
-  'schedule.retry': 'Opnieuw proberen',
-  'schedule.noClasses': 'Er zijn nog geen lessen ingepland.',
-  
-  // Locations
-  'locations.title': 'Onze Trainingslocaties',
-  'locations.address': 'Adres',
-  'locations.noMap': 'Geen kaart beschikbaar',
-  'locations.noLocations': 'Geen locaties beschikbaar',
-  'locations.loading': 'Locaties laden...',
-  'locations.error': 'Kon de locaties niet laden. Probeer het later opnieuw.',
-  'locations.questionsTitle': 'Vragen over onze locaties?',
-  'locations.postalAddress': 'Postadres',
-  'locations.phone': 'Telefoon',
-  'locations.email': 'Email',
-  'locations.hours': 'Openingstijden',
-  'locations.hoursText': 'Afhankelijk van lesrooster',
-  'locations.contactTitle': 'Vragen of Contact Nodig?',
-  'locations.contactDescription': 'Neem contact met ons op voor meer informatie over onze trainingslocaties, openingstijden of andere vragen.',
-  'locations.contactButton': 'Ga naar Contactpagina',
-  'locations.openMaps': 'Open in Google Maps',
-
-  // Contact
-  'contact.title': 'Neem Contact Op',
-  'contact.description': 'Heeft u vragen? Wij staan voor u klaar! Vul het formulier in of neem direct contact met ons op.',
-  'contact.info': 'Contactgegevens',
-  'contact.address': 'Postadres',
-  'contact.phone': 'Telefoon',
-  'contact.email': 'Email',
-  'contact.hours': 'Trainingsuren',
-  'contact.hoursText': 'Maandag, Woensdag, Donderdag en Zaterdag',
-  'contact.fullScheduleText': 'Bekijk het volledige rooster op de',
-  'contact.scheduleLink': 'rooster pagina',
-  'contact.formTitle': 'Stuur een Bericht',
-  'contact.name': 'Naam',
-  'contact.subject': 'Onderwerp',
-  'contact.message': 'Bericht',
-  'contact.send': 'Verstuur Bericht',
-  'contact.sending': 'Verzenden...',
-  'contact.error': 'Er ging iets mis bij het verzenden. Probeer het opnieuw.',
-  'contact.captchaRequired': 'Voltooi de CAPTCHA verificatie voordat u het formulier verstuurt.',
-  'common.captchaLabel': 'Ik ben geen robot, maar een Judoka',
-  'contact.success': 'Bericht Ontvangen!',
-  'contact.successText': 'Bedankt voor uw bericht! Wij nemen zo spoedig mogelijk contact met u op.',
-  'contact.subjectOptions.proefles': 'Gratis Proefles',
-  'contact.subjectOptions.inschrijving': 'Inschrijving',
-  'contact.subjectOptions.vraag': 'Algemene Vraag',
-  'contact.subjectOptions.anders': 'Anders',
-  'contact.subjectSelect': 'Selecteer een onderwerp',
-  
-  // Team
-  'team.title': 'Ons Team',
-  'team.description': 'Ontmoet onze ervaren en gepassioneerde trainers die zich inzetten voor uw ontwikkeling in judo.',
-  'team.cta': 'Interesse in Training?',
-  'team.ctaText': 'Kom langs voor een gratis proefles en maak kennis met ons team!',
-  'team.button': 'Boek een Gratis Proefles',
-  'team.error': 'Er is een fout opgetreden bij het laden van het team.',
-  'team.noInstructors': 'Geen instructeurs gevonden.',
-  'team.qualifications': 'Certificeringen & Prestaties:',
-  
-  // History
-  'history.title': 'Onze Rijke Geschiedenis',
-  
-  // Sectie 1: De Oprichting
-  'history.section1.title': 'De Oprichting (1950)',
-  'history.section1.p1': 'In november 1950 werd JUDOVERENIGING SHI-SEI opgericht. Een aantal enthousiaste judoka\'s van Johan v.d. Bruggen uit de Zoutmanstraat wilden een eigen club. Een club waar men goedkoper en ook meer kon trainen.',
-  'history.section1.p2': 'Onder leiding van Jacques Brakel, Hans van Diggelen en nog enkele anderen, werd op 5 november 1950 de Judovereniging SHI-SEI geboren, aangemeld bij de Kamer van Koophandel en bij de Nederlandse Amateur Judo Associatie aangesloten.',
-  'history.section1.p3': 'In de beginperiode werd geoefend in een gymzaal aan de Daal en Bergselaan, maar al vrij snel verhuisde men naar een permanente dojo in de Jan Hendrikstraat 9. Hier kon naar hartelust worden geoefend en zorgde de NAJA-bondstrainer Tokio Hirano, 6e dan, voor verbetering van de technieken. Er kon nu 7 dagen in de week worden getraind en desnoods de gehele dag. Het duurde dan ook niet lang of de trainingen brachten resultaat.',
-  'history.section1.tokioHiranoLink': 'https://judoinfo.com/hirano/#:~:text=Tokio%20Hirano:%20The%20Man%20Who%20Revolutionized%20Judo%20%7C%20Judo%20Info.%20Judo%20Info.%20Online%20Dojo.',
-  'history.section1.p4': 'Wil Wagner was de eerste zwartebandhouder van de club. Op een judo-zomerschool in Zwitserland ontving hij uit handen van grootmeester Hanno Rhi de fel begeerde band. Kort daarna behaalde Jan v.d. Toorn als tweede ook de 1e dan graad (zwarte band). Zij beiden waren in de beginperiode de technische leiding op de mat, aangevuld met bruine banden die fungeerden als oefenmeester op de diverse avonden.',
-  'history.section1.p5': 'De organisatorische leiding onder aanvoering van voorzitter Jacques Brakel zorgde dat de club ook in dat opzicht uitstekend draaide. Zwarte- en bruine banden werden ingeschreven voor Oefenmeester- en scheidsrechterscursussen en zorgden op die wijze voor de nodige "know-how".',
-  
-  // Highlight Box: Groei in de Jaren Zestig
-  'history.growth.title': 'Groei in de Jaren Zestig',
-  'history.growth.description': 'In de jaren zestig was SHI-SEI een toonaangevende club in Den Haag en een geduchte tegenstander op de mat. Steeds groter werd het ledenaantal en ook het kader breidde zich al maar uit.',
-  'history.growth.didYouKnow': 'Wist je dat SHI-SEI de moederclub werd van vele Haagse judoclubs zoals Lu-Gia-Jen, Guan-Tsui, Huan-Pi, De Doorkruiers, Guan Lin, Kagawa en Sumeju? Wil Wagner was de grondlegger van de meeste van deze clubs.',
-
-  // Sectie 2: Succesvolle Judoka's
-  'history.section2.title': 'Succesvolle Judoka\'s',
-  'history.section2.p1': 'Honderden zwartebandhouders heeft de club inmiddels afgeleverd waaronder een aantal bijzonder goede wedstrijdjudoka zoals Wil Tap, Jan Hendriks, Jaap Zoutendijk, Jan Dotinga, Kees Toonen, Koos Tap, Koos Letterie, Wim Lut, Giel Groos, Beb de Waal, Nico Schoemaker, Koos v.d. Mark, Wil Wagner, Jan v.d. Toorn, Ed Beekink, Wim de Rooy, Barry Oudshoorn, Henk Ammerlaan, John Lut en nog vele anderen.',
-
-  // Sectie 3: Wederopstanding & Nieuwe Locaties
-  'history.section3.title': 'Wederopstanding & Nieuwe Locaties',
-  'history.section3.p1': 'In 1970 ging de vereniging bijna ter ziele omdat een aantal leden op een buitengewone ledenvergadering het voltallige bestuur aan de kant zette. Dit liep uit op een falicante mislukking.',
-  'history.section3.p2': 'Doch in 1972 werd onder de bezielende leiding van Wim Lut een nieuw bestuur gevormd en werd vooral door vestigingen in het Westland de club weer nieuw leven ingeblazen. Dit had echter tot gevolg dat de accommodatie veel te klein werd en niet meer voldeed aan de eisen van die tijd.',
-  'history.section3.p3': 'Eindelijk in 1981 kreeg men door toedoen van wethouders Piet Vink en Adrie Duivestein de beschikking over een ruime accommodatie in de oude school aan de Nieuwe Schoolstraat 22-B.',
-  'history.section3.p4': 'Hier hebben we de beschikking over een ruime oefenzaal, twee ruime douche-gelegenheden, kantoor, Kantine, kleine oefenzaal en heel veel bergruimte. In deze ruimte hebben we nu niet alleen judo maar ook al sinds 1990 een taekwondo-afdeling met diverse eigen-kweek 1e en 2e dans. Ook een dart-afdeling zorgt voor de nodige gezelligheid.',
-
-  // Sectie 4: Verhuizingen & Recente Historie
-  'history.section4.title': 'Verhuizingen & Recente Historie',
-  'history.section4.p1': 'Helaas kwam in 2006 aan onze mooie eigen locatie een einde toen de gemeente besloot om ons huur contract op te zeggen om een drietal luxe woningen te bouwen op de plek van onze prachtige locatie. We konden helaas geen eigen nieuwe locatie vinden die wij konden bekostigen.',
-  'history.section4.p2': 'Wij besloten gebruik te maken van een buurthuis in de Fluitenbergstraat te Den Haag en een gymzaal aan de Beresteinlaan te Den Haag. In 2011 werd ontmoetingscentrum Morgenstond onze nieuwe locatie en werd de Gymzaal aan de Beresteinlaan wederom noodgedwongen ingeruild voor een mooiere gymzaal aan de Pachtersdreef.',
-  'history.section4.p3': 'Vanaf 2011 tot aan de corona crisis in 2020 gaven wij 6 dagen per week judo en taekwondo lessen. Ons ledental was gestaagd gegroeid naar ongeveer 140 judoleden en 40 taekwondoleden.',
-  'history.section4.p4': 'Vanaf 16 maart 2020 waren wij noodgedwongen te sluiten voor een korte periode i.v.m. met de corona virus infectie. Maar inmiddels zijn we weer op volle toeren hardop aan het trainen.',
-  
-  'history.timeline.title': 'Tijdlijn',
-  'history.stats.title': 'Vandaag de dag',
-  'history.stats.members': 'Leden',
-  'history.stats.region': 'Locatie',
-  'history.stats.city': 'Den Haag',
-  
-  'history.milestones.1950.title': 'Oprichting',
-  'history.milestones.1950.description': 'Start aan de Daal en Bergselaan.',
-  'history.milestones.1960.title': 'Gouden Jaren',
-  'history.milestones.1960.description': 'Moederclub van vele Haagse verenigingen.',
-  'history.milestones.1980.title': 'Nieuwe Schoolstraat',
-  'history.milestones.1980.description': 'Verhuizing naar de iconische dojo (1981).',
-  'history.milestones.2011.title': 'Morgenstond',
-  'history.milestones.2011.description': 'Nieuwe locatie aan de Pachtersdreef.',
-  'history.milestones.2025.title': 'Heden',
-  'history.milestones.2025.description': 'Nog steeds actief in Den Haag met 150+ leden.',
-  
-  // Pricing
-  'pricing.title': 'Lidmaatschap & Tarieven',
-  'pricing.popular': 'Populair',
-  'pricing.month': 'per maand',
-  
-  // Plans
-  'pricing.plans.youth.name': 'Jeugd (t/m 18 jaar)',
-  'pricing.plans.youth.price': '€27,50',
-  'pricing.plans.youth.f1': 'Onbeperkt trainen',
-  'pricing.plans.youth.f2': 'Toegang tot alle jeugdlessen',
-  'pricing.plans.youth.f3': 'Examen mogelijkheden',
-  'pricing.plans.youth.f4': 'Gratis judo pak t/m 17 jaar',
-  
-  'pricing.plans.adults.name': 'Volwassenen',
-  'pricing.plans.adults.price': '€27,50',
-  'pricing.plans.adults.f1': '2x per week trainen',
-  'pricing.plans.adults.f2': 'Toegang tot alle lessen',
-  'pricing.plans.adults.f3': 'Examen mogelijkheden',
-
-  'pricing.year': 'per jaar',
-  'pricing.plans.youth.yearly': '€330,-',
-  'pricing.plans.adults.yearly': '€330,-',
-
-  // Extra Info
-  'pricing.registrationFee': 'Eenmalige Inschrijfkosten',
-  'pricing.registrationFeeDescription': 'Deze kosten betaal je alleen bij je eerste inschrijving en gelden voor je gehele lidmaatschap.',
-  'pricing.ooievaarspas': 'Met de ooievaarspas tot 100% korting op de contributie mogelijk',
-  'pricing.startNow': 'Start Nu',
-  'pricing.loading': 'Tarieven laden...',
-  'pricing.error': 'Fout bij het laden van tarieven',
-  'pricing.noPrices': 'Geen prijsplannen beschikbaar',
-  'pricing.cta.title': 'Gratis Proefles',
-  'pricing.cta.desc': 'Probeer eerst een gratis proefles voordat u zich inschrijft. Geen verplichtingen!',
-  'pricing.cta.button': 'Boek een Gratis Proefles',
-  
-  // Exam
-  'exam.title': 'Examen Eisen',
-  "exam.participationTitle": "Deelname is altijd gewenst",
-  "exam.participationText": "Hoewel er voor volledige bandgraden specifieke minimale eisen gelden, zijn er tijdens de examens ook altijd tussentijdse slippen te behalen. Wij moedigen daarom iedereen aan om mee te doen.",
-  'exam.loading': 'Laden...',
-  'exam.error': 'Fout bij het laden van examen eisen',
-  'exam.noGrades': 'Er zijn geen examen eisen beschikbaar.',
-  'exam.download': 'Download Examen Document',
-  'exam.preview': 'Voorbeeld',
-  'exam.supplementary': 'Aanvullende Documenten',
-  'exam.minAge': 'Minimale leeftijd',
-
-  // Agenda
-  'agenda.badge': 'Jaarkalender',
-  'agenda.title': 'Jaaroverzicht Agenda',
-  'agenda.loading': 'Laden...',
-  'agenda.error': 'Fout bij het laden van de agenda',
-  'agenda.noEvents': 'Er zijn geen evenementen gepland.',
-  'agenda.vacation': 'Vakantie',
-  'agenda.holiday': 'Feestdag',
-  'agenda.exam': 'Examen',
-  'agenda.competition': 'Wedstrijd',
-  'agenda.other': 'Evenement',
-  'agenda.allDay': 'Hele dag',
-  'agenda.multiDay': 'Meerdaagse',
-  'agenda.registrationRequired': 'Aanmelden verplicht',
-  'agenda.registrationDeadline': 'Deadline',
-  'agenda.maxParticipants': 'Max deelnemers',
-  'agenda.moreInfo': 'Meer informatie',
-  'agenda.location': 'Locatie',
-
-  // Media/Albums
-  'media.title': 'Albums',
-  'media.description': 'Bekijk foto\'s en videos van trainingen, evenementen, wedstrijden en bijzondere momenten.',
-  'media.search': 'Zoek media albums...',
-  'media.loading': 'Albums laden...',
-  'media.error': 'Fout bij het laden van de albums',
-  'media.noAlbums': 'Er zijn nog geen albums beschikbaar.',
-  'media.viewAlbum': 'Bekijk Album',
-  'media.close': 'Sluiten',
-  'media.previous': 'Vorige',
-  'media.next': 'Volgende',
-  'media.filterContent': 'Alle Albums',
-  'media.filterPhotos': 'Alleen Foto\'s',
-  'media.filterVideos': 'Met Video\'s',
-  'media.download': 'Downloaden',
-  'media.downloadAll': 'Alles downloaden',
-  'media.downloadingAll': 'Downloaden...',
-
-  // News month filter
-  'news.filterMonth': 'Alle maanden',
-  'news.month.1': 'Januari',
-  'news.month.2': 'Februari',
-  'news.month.3': 'Maart',
-  'news.month.4': 'April',
-  'news.month.5': 'Mei',
-  'news.month.6': 'Juni',
-  'news.month.7': 'Juli',
-  'news.month.8': 'Augustus',
-  'news.month.9': 'September',
-  'news.month.10': 'Oktober',
-  'news.month.11': 'November',
-  'news.month.12': 'December',
-
-  // Regels (VCP en Regels)
-  'regels.title': 'VCP en Reglementen ',
-  'regels.description': 'Hier vindt u alle reglementen en huishoudelijke regels van Shi-Sei Sport. Lees deze documenten zorgvuldig door om op de hoogte te blijven van de regels en voorschriften.',
-  'regels.infoTitle': 'Belangrijke Informatie',
-  'regels.infoText': 'Alle leden en ouders/verzorgers worden geacht op de hoogte te zijn van deze reglementen. Klik op een document om meer details te lezen en gebruik de downloadknop om het PDF-bestand te downloaden.',
-  'regels.download': 'Download PDF',
-  'regels.noDocuments': 'Er zijn momenteel geen documenten beschikbaar.',
-  'regels.error': 'Kon de documenten niet laden. Probeer het later opnieuw.',
-
-  // VCP
-  'vcp.title': 'Vertrouwenscontactpersoon (VCP)',
-  'vcp.subtitle': 'Veilig en fijn sporten',
-  'vcp.contact': 'Contact',
-  'vcp.vcpName': 'VCP Naam',
-  'vcp.email': 'Email',
-  'vcp.whatDoesVcpDo': 'Wat doet de VCP?',
-  'vcp.forWhom': 'Voor wie is de VCP er?',
-  'vcp.whyContact': 'Waarom contact opnemen?',
-  'vcp.aboutVcp': 'Over de VCP',
-  'vcp.preventivePolicy': 'Preventief Beleid',
-  'vcp.crossingBehavior': 'Wat is grensoverschrijdend gedrag?',
-  'vcp.vcpTasks': 'Taken Vertrouwenscontactpersoon',
-
-  // Inschrijven (Enrollment)
-  'inschrijven.title': 'Inschrijven',
-  'inschrijven.description': 'Word lid van Shi-Sei Sport. Kies hieronder hoe u zich wilt inschrijven.',
-  'inschrijven.downloadOption': 'PDF Downloaden',
-  'inschrijven.downloadOptionDesc': 'Download het inschrijfformulier als PDF, vul het in en lever het in bij de club.',
-  'inschrijven.onlineOption': 'Online Inschrijven',
-  'inschrijven.onlineOptionDesc': 'Vul het formulier direct online in en ontvang automatisch een ingevuld PDF document.',
-  'inschrijven.formTitle': 'Online Inschrijfformulier',
-  'inschrijven.formDescription': 'Vul onderstaand formulier in en ontvang automatisch een ingevuld PDF document.',
-  'inschrijven.infoTitle': 'Hoe werkt het?',
-  'inschrijven.step1': 'Download het formulier hieronder',
-  'inschrijven.step2': 'Vul het volledig in',
-  'inschrijven.step3': 'Lever het in bij de club of stuur het digitaal op',
-  'inschrijven.download': 'Download',
-  'inschrijven.noDocuments': 'Er zijn momenteel geen inschrijfformulieren beschikbaar.',
-  'inschrijven.error': 'Kon de formulieren niet laden. Probeer het later opnieuw.',
-  'inschrijven.ctaTitle': 'Hulp nodig bij inschrijven?',
-  'inschrijven.ctaText': 'Neem contact met ons op als u vragen heeft over het inschrijfproces of hulp nodig heeft bij het invullen van de formulieren.',
-  'inschrijven.ctaButton': 'Neem Contact Op',
-  'inschrijven.heroTagline': 'Begin jouw judoreis vandaag',
-  'inschrijven.heroSub': 'Word lid van onze club en ontdek de kracht van judo',
-
-  // Enrollment Form
-  'enrollment.form.personalInfo': 'Persoonlijke Gegevens',
-  'enrollment.form.firstName': 'Roepnaam',
-  'enrollment.form.middleName': 'Tussenvoegsel',
-  'enrollment.form.lastName': 'Achternaam',
-  'enrollment.form.dateOfBirth': 'Geboortedatum',
-  'enrollment.form.email': 'E-mail',
-  'enrollment.form.phone': 'Telefoonnummer',
-  'enrollment.form.address': 'Adres',
-  'enrollment.form.street': 'Straatnaam',
-  'enrollment.form.houseNumber': 'Huisnummer',
-  'enrollment.form.postalCode': 'Postcode',
-  'enrollment.form.city': 'Plaats',
-  'enrollment.form.emergencyContact': 'Noodcontactpersoon',
-  'enrollment.form.contactName': 'Naam',
-  'enrollment.form.contactPhone': 'Telefoonnummer',
-  'enrollment.form.contactRelation': 'Relatie',
-  'enrollment.form.relationPlaceholder': 'Bijv. ouder, partner, vriend',
-  'enrollment.form.guardianName': 'Naam ouder/voogd',
-  'enrollment.form.guardianNamePlaceholder': 'Alleen invullen bij inschrijving minderjarige',
-  'enrollment.form.judoExperience': 'Judo Ervaring',
-  'enrollment.form.experience': 'Ervaringsniveau',
-  'enrollment.form.experienceBeginner': 'Beginner',
-  'enrollment.form.experienceSome': 'Enige ervaring',
-  'enrollment.form.experienceAdvanced': 'Gevorderd',
-  'enrollment.form.judoGrade': 'Judo Graad',
-  'enrollment.form.judoGrade.none': 'Geen / Onbekend',
-  'enrollment.form.judoGrade.6kyu': '6e kyu (Wit)',
-  'enrollment.form.judoGrade.5kyu': '5e kyu (Geel)',
-  'enrollment.form.judoGrade.4kyu': '4e kyu (Oranje)',
-  'enrollment.form.judoGrade.3kyu': '3e kyu (Groen)',
-  'enrollment.form.judoGrade.2kyu': '2e kyu (Blauw)',
-  'enrollment.form.judoGrade.1kyu': '1e kyu (Bruin)',
-  'enrollment.form.judoGrade.dan': 'Dan (Zwarte band)',
-  'enrollment.form.preferredDays': 'Voorkeur Trainingsdagen',
-  'enrollment.form.day.monday': 'Maandag',
-  'enrollment.form.day.wednesday': 'Woensdag',
-  'enrollment.form.day.thursday': 'Donderdag',
-  'enrollment.form.day.saturday': 'Zaterdag',
-  'enrollment.form.medicalInfo': 'Medische Informatie',
-  'enrollment.form.medicalInfoPlaceholder': 'Allergieën, medicatie, beperkingen...',
-  'enrollment.form.remarks': 'Opmerkingen',
-  'enrollment.form.remarksPlaceholder': 'Eventuele vragen of opmerkingen...',
-  'enrollment.form.paymentMethod': 'Betaalmethode',
-  'enrollment.form.paymentMethodNote': 'Selecteer uw betaalmethode. Met de Ooievaarspas ontvangt u korting en hoeft u geen machtiging in te vullen.',
-  'enrollment.form.ooievaarspas': 'Ik heb een Ooievaarspas',
-  'enrollment.form.ooievaarspasNote': 'Vink aan als u een geldige Ooievaarspas heeft.',
-  'enrollment.form.ooievaarspasNumber': 'Ooievaarspas nummer',
-  'enrollment.form.ooievaarspasNumberPlaceholder': 'Vul uw Ooievaarspas nummer in',
-  'enrollment.form.ooievaarspasCheckLocation': 'Uw Ooievaarspas wordt op locatie gecontroleerd.',
-  'enrollment.form.machtiging': 'Machtiging (Automatische Incasso)',
-  'enrollment.form.machtigingNote': 'Voor automatische incasso hebben we uw bankgegevens nodig.',
-  'enrollment.form.accountHolder': 'Naam Rekeninghouder',
-  'enrollment.form.iban': 'IBAN Rekeningnummer',
-  'enrollment.form.submit': 'Inschrijving Versturen',
-  'enrollment.form.submitting': 'Bezig met versturen...',
-  'enrollment.form.success': 'Inschrijving Ontvangen!',
-  'enrollment.form.successText': 'Bedankt voor uw inschrijving! We hebben uw gegevens ontvangen en nemen binnen 3 werkdagen contact met u op.',
-  'enrollment.form.pdfGenerated': 'Een ingevuld PDF formulier is automatisch gegenereerd en opgeslagen.',
-  'enrollment.form.error': 'Er ging iets mis bij het versturen. Probeer het opnieuw.',
-  'enrollment.form.captchaRequired': 'Voltooi de CAPTCHA verificatie voordat u het formulier verstuurt.',
-  'enrollment.form.agreeTermsPre': 'Ik ga akkoord met de',
-  'enrollment.form.agreeTermsLink': 'Algemene Voorwaarden',
-  'enrollment.form.agreeTermsMid': 'en het',
-  'enrollment.form.agreePrivacyLink': 'Privacy Beleid',
-  'enrollment.form.agreeTermsPost': 'van Shi-Sei Sport.',
-  'enrollment.form.signature': 'Handtekening',
-  'enrollment.form.signatureClear': 'Wissen',
-  'enrollment.form.disclaimer': 'Door dit formulier in te dienen ga je akkoord met ons privacy beleid.',
-
-  // Enrollment Confirmation
-  'enrollment.confirm.title': 'Gegevens Bevestigen',
-  'enrollment.confirm.description': 'Controleer uw gegevens voordat u de inschrijving verstuurt.',
-  'enrollment.confirm.cancel': 'Terug',
-  'enrollment.confirm.submit': 'Bevestigen & Versturen',
-
-  // Trial
-  'trial.title': 'Boek een Gratis Proefles',
-  'trial.description': 'Probeer judo bij Shi-Sei Sport. Geen verplichtingen, gewoon komen en ervaren!',
-  'trial.benefits': 'Wat krijg je bij een proefles?',
-  'trial.benefit1': 'Gratis en zonder verplichtingen',
-  'trial.benefit2': 'Kennis maken met onze trainers',
-  'trial.benefit3': 'Ervaren van onze trainingsstijl',
-  'trial.benefit4': 'Informatie over lidmaatschap',
-  'trial.benefit5': 'Direct kunnen starten met trainen',
-  'trial.days': 'Trainingsdagen',
-  'trial.daysText': 'Maandag, Woensdag, Donderdag en Zaterdag',
-  'trial.scheduleLink': 'Bekijk volledige rooster',
-  'trial.bring': 'Wat mee te nemen?',
-  'trial.bring1': 'Comfortabele sportkleding',
-  'trial.bring2': 'Waterfles',
-  'trial.bring3': 'Slippers',
-  'trial.bring4': 'Judo-gi wordt de eerste keer geleend',
-  'trial.forWho': 'Voor wie?',
-  'trial.forWhoText': 'Iedereen vanaf 4 jaar is welkom! Of je nu beginner bent of al ervaring hebt, we hebben lessen voor alle niveaus.',
-  'trial.formTitle': 'Boek je Proefles',
-  'trial.age': 'Leeftijd',
-  'trial.experience': 'Ervaring',
-  'trial.experienceOptions.beginner': 'Beginner',
-  'trial.experienceOptions.some': 'Enige ervaring',
-  'trial.experienceOptions.advanced': 'Gevorderd',
-  'trial.preferredDay': 'Voorkeur trainingsdag',
-  'trial.preferredDayOptions.maandag': 'Maandag',
-  'trial.preferredDayOptions.woensdag': 'Woensdag',
-  'trial.preferredDayOptions.donderdag': 'Donderdag',
-  'trial.preferredDayOptions.zaterdag': 'Zaterdag',
-  'trial.preferredDayOptions.none': 'Geen voorkeur',
-  'trial.preferredDayAdultHint': 'Training voor volwassenen (18+) is alleen op maandag en donderdag.',
-  'trial.message': 'Opmerkingen (optioneel)',
-  'trial.messagePlaceholder': 'Heb je vragen of specifieke wensen?',
-  'trial.button': 'Boek Gratis Proefles',
-  'trial.submitting': 'Verzenden...',
-  'trial.error': 'Er is een fout opgetreden bij het verzenden. Probeer het later opnieuw.',
-  'trial.captchaRequired': 'Vul alstublieft de CAPTCHA in om door te gaan.',
-  'trial.success': 'Bedankt voor je aanmelding!',
-  'trial.successText': 'We hebben je aanvraag ontvangen en nemen binnen 3 dagen contact met je op om een geschikt tijdstip te plannen.',
-  'trial.disclaimer': 'Door te boeken ga je akkoord met onze privacy policy. We nemen binnen 3 dagen contact met je op.',
-  
-  // Footer
-  'footer.club': 'Shi-Sei Sport',
-  'footer.description': 'De oudste judo club van Den Haag. Judo lessen sinds 1950.',
-  'footer.quickLinks': 'Snelle Links',
-  'footer.contact': 'Contact',
-  'footer.contactText': 'Voor vragen of meer informatie, neem contact met ons op via de contactpagina.',
-  'footer.jbn': 'Officieel aangesloten bij de JBN.',
-  'footer.partners': 'Partners',
-  'footer.ooievaarspas': 'Wij accepteren de Ooievaarspas.',
-  'footer.copyright': 'Alle rechten voorbehouden.',
-  'footer.followUs': 'Volg ons op',
-  'footer.privacy': 'Privacy Policy',
-  'footer.rules': 'Huishoudelijke Regels',
-  'footer.terms': 'Algemene Voorwaarden',
-
-  // Privacy Policy
-  'privacy.title': 'Privacy Policy',
-  'privacy.lastUpdated': 'Laatst bijgewerkt',
-  'privacy.intro.title': 'Inleiding',
-  'privacy.intro.text1': 'Shi-Sei Sport respecteert de privacy van alle bezoekers van onze website en leden van onze judoclub. In deze privacy policy leggen we uit hoe we omgaan met uw persoonsgegevens.',
-  'privacy.intro.text2': 'Deze privacy policy is van toepassing op het gebruik van de website en de diensten die worden aangeboden door Shi-Sei Sport.',
-  'privacy.collection.title': 'Welke gegevens verzamelen wij?',
-  'privacy.collection.intro': 'Wij verzamelen verschillende soorten gegevens voor verschillende doeleinden:',
-  'privacy.collection.item1': 'Contactgegevens zoals naam, e-mailadres en telefoonnummer',
-  'privacy.collection.item2': 'Leeftijd en geboortedatum (voor juiste indeling in trainingsgroepen)',
-  'privacy.collection.item3': 'Ervaring en bandniveau (voor juiste trainingsgroepen)',
-  'privacy.collection.item4': 'Communicatie via onze contactformulieren en e-mail',
-  'privacy.collection.item5': 'Foto- en beeldmateriaal (voor promotionele doeleinden zoals de website en sociale media)',
-  'privacy.usage.title': 'Hoe gebruiken wij uw gegevens?',
-  'privacy.usage.intro': 'Wij gebruiken uw persoonsgegevens voor de volgende doeleinden:',
-  'privacy.usage.item1': 'Het verwerken van uw aanmelding voor lidmaatschap of proefles',
-  'privacy.usage.item2': 'Het versturen van informatie over trainingen, evenementen en examens',
-  'privacy.usage.item3': 'Het beantwoorden van vragen en verzoeken',
-  'privacy.usage.item4': 'Het verbeteren van onze diensten en website',
-  'privacy.protection.title': 'Hoe beschermen wij uw gegevens?',
-  'privacy.protection.text1': 'Wij nemen de bescherming van uw gegevens serieus en nemen passende beveiligingsmaatregelen om misbruik, verlies en ongeautoriseerde toegang te voorkomen.',
-  'privacy.protection.text2': 'Uw gegevens worden opgeslagen op beveiligde servers en zijn alleen toegankelijk voor geautoriseerd personeel van Shi-Sei Sport.',
-  'privacy.rights.title': 'Uw rechten',
-  'privacy.rights.intro': 'U heeft de volgende rechten met betrekking tot uw persoonsgegevens:',
-  'privacy.rights.item1': 'Recht op inzage: u kunt opvragen welke gegevens wij van u verwerken',
-  'privacy.rights.item2': 'Recht op correctie: u kunt vragen om uw gegevens te corrigeren als deze onjuist zijn',
-  'privacy.rights.item3': 'Recht op verwijdering: u kunt vragen om uw gegevens te verwijderen',
-  'privacy.rights.item4': 'Recht op bezwaar: u kunt bezwaar maken tegen de verwerking van uw gegevens',
-  'privacy.cookies.title': 'Cookies',
-  'privacy.cookies.text1': 'Onze website gebruikt alleen functionele opslag die noodzakelijk is voor het goed functioneren van de website. We gebruiken geen tracking cookies of analytics die uw surfgedrag volgen.',
-  'privacy.cookies.intro': 'De volgende gegevens worden lokaal opgeslagen in uw browser:',
-  'privacy.cookies.item1': 'Taalvoorkeur (localStorage) — onthoudt of u de Nederlandse of Engelse versie van de site heeft gekozen.',
-  'privacy.cookies.item2': 'Donkere modus (localStorage) — onthoudt of u het lichte of donkere thema heeft gekozen.',
-  'privacy.contact.title': 'Contact',
-  'privacy.contact.text': 'Heeft u vragen over onze privacy policy of wilt u gebruik maken van uw rechten? Neem dan contact met ons op:',
-  'privacy.contact.email': 'E-mail',
-
-  // Terms & Conditions
-  'terms.title': 'Algemene Voorwaarden',
-  'terms.lastUpdated': 'Laatst bijgewerkt',
-  'terms.intro.title': 'Algemene Bepalingen',
-  'terms.intro.text': 'Deze algemene voorwaarden zijn van toepassing op het lidmaatschap van Shi-Sei Sport. Door lid te worden accepteert u deze voorwaarden.',
-
-  'terms.expect.title': 'Wat kan je van Shi-Sei Sport verwachten',
-  'terms.expect.item1': 'Verantwoord sporten in een veilige sportomgeving.',
-  'terms.expect.item2': 'Uw privé- en adresgegevens worden niet zonder overleg met u gedeeld met derden.',
-  'terms.expect.item3': 'De lessen worden gegeven door vakbekwame, opgeleide en/of enthousiaste trainers.',
-  'terms.expect.item4': 'Begeleiders, vrijwilligers en/of trainers zijn allen in het bezit van een geldige VOG-verklaring omtrent gedrag.',
-  'terms.expect.item5': 'Iedere afdeling heeft zijn eigen en vaste professionele trainer/begeleider.',
-  'terms.expect.item6': 'Professioneel werkende sportclub met voldoende aanwezige EHBO-middelen.',
-  'terms.expect.item7': 'Een grote dosis enthousiasme van iedere leidinggevende.',
-
-  'terms.enrollment.title': 'Inschrijven',
-  'terms.enrollment.item1': 'Lees voor het invullen en inleveren van het inschrijfformulier goed de algemene voorwaarden.',
-  'terms.enrollment.item2pre': 'Ieder lid is verplicht kennis te nemen van het',
-  'terms.enrollment.regelementLink': 'algemeen reglement',
-  'terms.enrollment.huishoudelijkLink': 'huishoudelijk reglement',
-  'terms.enrollment.item2mid': 'en de',
-  'terms.enrollment.gedragscodeLink': 'gedragscode',
-
-  'terms.signature.title': 'Ondertekening',
-  'terms.signature.item1': 'Door het aanvinken of het ondertekenen van het online inschrijfformulier verklaart ondergetekende zich akkoord met de lidmaatschapsvoorwaarden van Shi-Sei Sport, waarvan de Algemene voorwaarden en de dojo-regels een onverbrekelijk onderdeel zijn, en begrijpt wat hierin vermeld wordt.',
-  'terms.signature.item2': 'De Algemene voorwaarden zijn zichtbaar op de website www.shi-sei.nl.',
-  'terms.signature.item3': 'De rechten uit hoofde van het lidmaatschap zijn strikt persoonlijk.',
-  'terms.signature.item4': 'Jongeren onder de 18 jaar kunnen uitsluitend worden ingeschreven door de ouder(s) of wettelijke vertegenwoordigers.',
-  'terms.signature.item5': 'Door ondertekening van het machtigingsformulier verleen je toestemming aan Shi-Sei Sport om tot wederopzegging automatisch de kosten van de door u opgegeven bankrekening af te schrijven, de bedragen welke verschuldigd zijn voor het vooruitbetalen van de lidmaatschapsbijdrage en andere gelden die Shi-Sei Sport toekomen.',
-
-  'terms.membership.title': 'Lidmaatschap',
-  'terms.membership.item1': 'Het lidmaatschap gaat in na ontvangst van de inschrijving en betaling van de contributie.',
-  'terms.membership.item2': 'Jongeren onder de 18 jaar kunnen uitsluitend worden ingeschreven door de ouder(s) of wettelijke vertegenwoordigers.',
-  'terms.membership.item3pre': 'Leden dienen zich te houden aan de',
-  'terms.membership.item3mid': 'en',
-  'terms.membership.item4': 'Het lidmaatschap is persoonlijk en niet overdraagbaar.',
-
-  'terms.link.huisregels': 'huisregels',
-  'terms.link.gedragscode': 'gedragscode',
-
-  'terms.rates.title': 'Tarieven',
-  'terms.rates.item1': 'Indien noodzakelijk kunnen tarieven door Shi-Sei Sport worden aangepast.',
-  'terms.rates.item2': 'Wijziging van het lidmaatschapstarief wordt 1 maand voordat het ingaat per e-mail bekendgemaakt.',
-  'terms.rates.item3': 'Indien een prijsverhoging plaatsvindt, heeft het lid het recht om de overeenkomst binnen 4 weken na de bekendmaking te ontbinden.',
-  'terms.rates.item4': 'Het eventueel vooruitbetaalde abonnementsgeld wordt terugbetaald.',
-
-  'terms.payment.title': 'Betaling',
-  'terms.payment.item1': 'Het lid of zijn vertegenwoordiger dient zorg te dragen voor een probleemloze betaling van de lidmaatschapsbijdragen.',
-  'terms.payment.item2': 'De lidmaatschapsbijdragen worden per kwartaal automatisch geïncasseerd rond de eerste van de nieuwe maand.',
-  'terms.payment.item3': 'De contributie wordt per kwartaal vooruitbetaald.',
-  'terms.payment.item4': 'Bij niet-tijdige betaling kan de toegang tot lessen worden opgeschort.',
-  'terms.payment.item5': 'Restitutie van contributie is niet mogelijk, behalve in uitzonderlijke gevallen.',
-
-  'terms.behavior.title': 'Gedrag en Huisregels',
-  'terms.behavior.item1': 'Leden dienen respectvol om te gaan met andere leden, trainers en materiaal.',
-  'terms.behavior.item2pre': 'Leden dienen zich te houden aan de',
-  'terms.behavior.item2mid': 'en',
-  'terms.behavior.item3': 'Het is niet toegestaan onder invloed van alcohol of drugs deel te nemen aan trainingen.',
-  'terms.behavior.item4': 'Bij ernstige overtredingen kan het bestuur besluiten tot schorsing of beëindiging van het lidmaatschap.',
-
-  'terms.exams.title': 'Examens en Wedstrijden',
-  'terms.exams.item1': 'De kosten voor de speciale wedstrijdgroep, wedstrijden en examens zijn niet bij de lesgelden inbegrepen.',
-  'terms.exams.item2': 'In het geval van afwezigheid van het lid bij een examen of wedstrijd zal er geen restitutie plaatsvinden van inschrijfgeld.',
-
-  'terms.liability.title': 'Eigen Risico',
-  'terms.liability.text1': 'Shi-Sei Sport is niet aansprakelijk voor letsel, verlies of schade die tijdens of als gevolg van sportactiviteiten ontstaat, tenzij dit het gevolg is van grove nalatigheid of opzet van de vereniging.',
-  'terms.liability.text2': 'Leden worden dringend geadviseerd om een eigen sportblessure- of ongevallenverzekering af te sluiten.',
-
-  'terms.photos.title': 'Foto\'s en Beeldmateriaal',
-  'terms.photos.item1': 'Het lid en andere bezoekers gaan akkoord met het gebruik van zijn of haar beeltenis in druk, op foto, video e.d. voor website en andere promotionele doeleinden zonder daarvoor vergoeding te claimen.',
-  'terms.photos.item2': 'Bij bezwaar hiertegen kunt u dit schriftelijk op het inschrijfformulier kenbaar maken en altijd wijzigingen per e-mail doorgeven.',
-
-  'terms.changes.title': 'Wijzigingen',
-  'terms.changes.text': 'Shi-Sei Sport behoudt zich het recht voor deze algemene voorwaarden te wijzigen. Leden worden hiervan op de hoogte gesteld via de website of per e-mail.',
-
-  'terms.cancellation.title': 'Opzegging',
-  'terms.cancellation.item1': 'Het lidmaatschap kan met een opzegtermijn van één maand worden beëindigd.',
-  'terms.cancellation.item2': 'Opzegging dient schriftelijk of per e-mail te gebeuren.',
-  'terms.cancellation.item3': 'Bij opzegging binnen de contractperiode blijft men de contributie verschuldigd tot het einde van die periode.',
-
-  'terms.contact.title': 'Vragen',
-  'terms.contact.text': 'Heeft u vragen over deze algemene voorwaarden? Neem dan contact met ons op:',
-  'terms.contact.email': 'E-mail',
-
-  // Common
-  'common.loading': 'Laden...',
-  'common.required': '*',
-  'common.select': 'Selecteer',
-  'common.optional': '(optioneel)',
-  'common.invalidEmail': 'Voer een geldig e-mailadres in, bijv. naam@voorbeeld.nl',
-  'common.invalidPhone': 'Voer een geldig telefoonnummer in, bijv. +31-612345678',
-  'common.invalidIban': 'Voer een geldig IBAN in, bijv. NL91 ABNA 0417 1643 00',
-  'common.invalidPostalCode': 'Voer een geldige postcode in, bijv. 1234 AB',
-  'common.invalidAge': 'Voer een geldige leeftijd in (4–100)',
-  'common.ibanCountryHint': 'Het formulier past zich automatisch aan op de landcode.',
-
-  // Placeholders
-  'placeholder.name': 'Uw volledige naam',
-  'placeholder.firstName': 'Roepnaam',
-  'placeholder.middleName': 'Tussenvoegsel (optioneel)',
-  'placeholder.lastName': 'Achternaam',
-  'placeholder.email': 'naam@voorbeeld.nl',
-  'placeholder.phone': '06-12345678',
-  'placeholder.street': 'Straatnaam',
-  'placeholder.houseNumber': 'Nr.',
-  'placeholder.city': 'Woonplaats',
-  'placeholder.accountHolder': 'Naam rekeninghouder',
-  'placeholder.age': 'bijv. 8',
-  'placeholder.message': 'Typ hier uw bericht...',
-
-  // 404
-  '404.title': 'Pagina niet gevonden',
-  '404.description': 'De pagina die u zoekt bestaat niet of is verplaatst.',
-  '404.home': 'Naar Homepage',
-  '404.back': 'Terug',
-};
-
-const translationsEn: Record<string, string> = {
-  // Navigation
-  'nav.home': 'Home',
-  'nav.about': 'About Us',
-  'nav.team': 'The Team',
-  'nav.history': 'History',
-  'nav.locations': 'Locations',
-  'nav.rules': 'Rules & Regulations',
-  'nav.info': 'Information',
-  'nav.schedule': 'Schedule',
-  'nav.pricing': 'Pricing',
-  'nav.exam': 'Exam Requirements',
-  'nav.enrollment': 'Enrollment',
-  'nav.agenda': 'Agenda',
-  'nav.news': 'News',
-  'nav.media': 'Media',
-  'nav.contact': 'Contact',
-  'nav.trial': 'Free Trial',
-  
-  // Hero
-  'hero.title': 'Judo',
-  'hero.subtitle': 'SHI-SEI SPORT',
-  'hero.description': 'The oldest judo club in The Hague',
-  'hero.text': 'We teach judo on Monday, Wednesday, Thursday and Saturday. Experience the power of discipline, respect and self-confidence in our dojo.',
-  'hero.button': 'Try a Trial Lesson',
-  'hero.scheduleButton': 'View the Schedule',
-  
-  // News
-  'news.latest': 'Latest News',
-  'news.all': 'All news',
-  'news.search': 'Search news...',
-  'news.readMore': 'Read more',
-  'news.clickToRead': 'Click to read the full article...',
-  'news.loading': 'Loading...',
-  'news.notFound': 'News not found',
-  'news.back': 'Back to news',
-  'news.backAll': 'Back to all news',
-  'news.noNews': 'No news available yet.',
-  'news.title': 'Latest News',
-  'news.description': 'Stay up to date with the latest news and events at Shi-Sei Sport.',
-  'news.share': 'Share this article',
-  'news.shareCopied': 'Link copied!',
-  'news.noContent': 'This article has no content.',
-
-  // Schedule
-  'schedule.title': 'Lessons & Schedule',
-  'schedule.description': 'We teach judo on Monday, Wednesday, Thursday and Saturday',
-  'schedule.loading': 'Loading schedule...',
-  'schedule.error': 'Could not load schedule. Please try again later.',
-  'schedule.retry': 'Try again',
-  'schedule.noClasses': 'No classes scheduled yet.',
-  
-  // Locations
-  'locations.title': 'Our Training Locations',
-  'locations.address': 'Address',
-  'locations.noMap': 'No map available',
-  'locations.noLocations': 'No locations available',
-  'locations.loading': 'Loading locations...',
-  'locations.error': 'Could not load locations. Please try again later.',
-  'locations.questionsTitle': 'Questions about our locations?',
-  'locations.postalAddress': 'Postal Address',
-  'locations.phone': 'Phone',
-  'locations.email': 'Email',
-  'locations.hours': 'Opening Hours',
-  'locations.hoursText': 'Depends on class schedule',
-  'locations.contactTitle': 'Questions or Need Contact?',
-  'locations.contactDescription': 'Contact us for more information about our training locations, opening hours, or other questions.',
-  'locations.contactButton': 'Go to Contact Page',
-  'locations.openMaps': 'Open in Google Maps',
-
-  // Contact
-  'contact.title': 'Contact Us',
-  'contact.description': 'Have questions? We are here for you! Fill out the form or contact us directly.',
-  'contact.info': 'Contact Information',
-  'contact.address': 'Postal Address',
-  'contact.phone': 'Phone',
-  'contact.email': 'Email',
-  'contact.hours': 'Training Hours',
-  'contact.hoursText': 'Monday, Wednesday, Thursday and Saturday',
-  'contact.fullScheduleText': 'View the full schedule on the',
-  'contact.scheduleLink': 'schedule page',
-  'contact.formTitle': 'Send a Message',
-  'contact.name': 'Name',
-  'contact.subject': 'Subject',
-  'contact.message': 'Message',
-  'contact.send': 'Send Message',
-  'contact.sending': 'Sending...',
-  'contact.error': 'Something went wrong while sending. Please try again.',
-  'contact.captchaRequired': 'Please complete the CAPTCHA verification before submitting the form.',
-  'common.captchaLabel': "I'm not a robot, but a Judoka",
-  'contact.success': 'Message Received!',
-  'contact.successText': 'Thank you for your message! We will contact you as soon as possible.',
-  'contact.subjectOptions.proefles': 'Free Trial',
-  'contact.subjectOptions.inschrijving': 'Registration',
-  'contact.subjectOptions.vraag': 'General Question',
-  'contact.subjectOptions.anders': 'Other',
-  'contact.subjectSelect': 'Select a subject',
-  
-  // Team
-  'team.title': 'Our Team',
-  'team.description': 'Meet our experienced and passionate trainers who are committed to your development in judo.',
-  'team.cta': 'Interested in Training?',
-  'team.ctaText': 'Come by for a free trial lesson and meet our team!',
-  'team.button': 'Book a Free Trial',
-  'team.error': 'An error occurred while loading the team.',
-  'team.noInstructors': 'No instructors found.',
-  'team.qualifications': 'Certifications & Achievements:',
-  
-  'history.title': 'Our Rich History',
-  
-  // Section 1: The Foundation
-  'history.section1.title': 'The Foundation (1950)',
-  'history.section1.p1': 'In November 1950, JUDO ASSOCIATION SHI-SEI was founded. A group of enthusiastic judokas from Johan v.d. Bruggen in the Zoutmanstraat wanted their own club. A place where they could train more often and at a lower cost.',
-  'history.section1.p2': 'Led by Jacques Brakel, Hans van Diggelen and several others, Judo Association SHI-SEI was born on November 5, 1950, registered with the Chamber of Commerce and affiliated with the Dutch Amateur Judo Association.',
-  'history.section1.p3': 'Initially, training took place in a gym on Daal en Bergselaan, but the club soon moved to a permanent dojo at Jan Hendrikstraat 9. Here, members could train to their heart\'s content, and NAJA national trainer Tokio Hirano, 6th dan, ensured the improvement of techniques. Training was now possible seven days a week, all day if necessary. It didn\'t take long for the training to yield results.',
-  'history.section1.tokioHiranoLink': 'https://judoinfo.com/hirano/#:~:text=Tokio%20Hirano:%20The%20Man%20Who%20Revolutionized%20Judo%20%7C%20Judo%20Info.%20Judo%20Info.%20Online%20Dojo.',
-  'history.section1.p4': 'Wil Wagner was the club\'s first black belt holder. At a judo summer school in Switzerland, he received the coveted belt from grandmaster Hanno Rhi. Shortly after, Jan v.d. Toorn also achieved the 1st dan grade (black belt) as the second member. Both of them provided the technical leadership on the mat during the early period, supplemented by brown belts who acted as instructors on various evenings.',
-  'history.section1.p5': 'The organizational leadership under chairman Jacques Brakel ensured the club ran excellently in that regard as well. Black and brown belts were enrolled in Instructor and referee courses, thus providing the necessary "know-how".',
-
-  // Highlight Box: Growth in the Sixties
-  'history.growth.title': 'Growth in the Sixties',
-  'history.growth.description': 'In the sixties, SHI-SEI was a leading club in The Hague and a formidable opponent on the mat. The membership grew steadily larger and the staff also expanded continuously.',
-  'history.growth.didYouKnow': 'Did you know that SHI-SEI became the mother club of many Hague judo clubs such as Lu-Gia-Jen, Guan-Tsui, Huan-Pi, De Doorkruiers, Guan Lin, Kagawa, and Sumeju? Wil Wagner was the founder of most of these clubs.',
-
-  // Section 2: Successful Judokas
-  'history.section2.title': 'Successful Judokas',
-  'history.section2.p1': 'The club has produced hundreds of black belt holders, including a number of particularly good competition judokas such as Wil Tap, Jan Hendriks, Jaap Zoutendijk, Jan Dotinga, Kees Toonen, Koos Tap, Koos Letterie, Wim Lut, Giel Groos, Beb de Waal, Nico Schoemaker, Koos v.d. Mark, Wil Wagner, Jan v.d. Toorn, Ed Beekink, Wim de Rooy, Barry Oudshoorn, Henk Ammerlaan, John Lut and many others.',
-
-  // Section 3: Resurrection & New Locations
-  'history.section3.title': 'Resurrection & New Locations',
-  'history.section3.p1': 'In 1970, the association almost ceased to exist when a number of members dismissed the entire board at an extraordinary general meeting. This turned out to be a disastrous failure.',
-  'history.section3.p2': 'However, in 1972 a new board was formed under the inspiring leadership of Wim Lut, and the club was given new life, particularly through branches in Westland. This, however, resulted in the accommodation becoming far too small and no longer meeting the requirements of the time.',
-  'history.section3.p3': 'Finally, in 1981, through aldermen Piet Vink and Adrie Duivestein, we obtained a spacious accommodation in the old school at Nieuwe Schoolstraat 22-B.',
-  'history.section3.p4': 'Here we have a spacious training hall, two spacious shower facilities, office, canteen, small training hall and plenty of storage space. In this space we now have not only judo but also, since 1990, a taekwondo department with various home-grown 1st and 2nd dans. A darts department also provides the necessary conviviality.',
-
-  // Section 4: Relocations & Recent History
-  'history.section4.title': 'Relocations & Recent History',
-  'history.section4.p1': 'Unfortunately, our beautiful own location came to an end in 2006 when the municipality decided to terminate our lease contract to build three luxury apartments on the site of our beautiful location. Unfortunately, we could not find our own new location that we could afford.',
-  'history.section4.p2': 'We decided to use a community center in Fluitenbergstraat in The Hague and a gym at Beresteinlaan in The Hague. In 2011, the Morgenstond meeting center became our new location and the gym at Beresteinlaan was once again necessarily exchanged for a nicer gym at Pachtersdreef.',
-  'history.section4.p3': 'From 2011 until the corona crisis in 2020, we taught judo and taekwondo 6 days a week. Our membership had grown steadily to approximately 140 judo members and 40 taekwondo members.',
-  'history.section4.p4': 'From March 16, 2020, we were forced to close for a short period due to the corona virus infection. But now we are back to training at full speed.',
-  
-  // Timeline & Stats
-  'history.timeline.title': 'Timeline',
-  'history.stats.title': 'Current Status',
-  'history.stats.members': 'Members',
-  'history.stats.region': 'Location',
-  'history.stats.city': 'The Hague',
-  
-  'history.milestones.1950.title': 'Foundation',
-  'history.milestones.1950.description': 'Start at Daal en Bergselaan.',
-  'history.milestones.1960.title': 'Golden Years',
-  'history.milestones.1960.description': 'Mother club of many Hague associations.',
-  'history.milestones.1980.title': 'Nieuwe Schoolstraat',
-  'history.milestones.1980.description': 'Relocation to the iconic dojo (1981).',
-  'history.milestones.2011.title': 'Morgenstond',
-  'history.milestones.2011.description': 'New location at Pachtersdreef.',
-  'history.milestones.2025.title': 'Present',
-  'history.milestones.2025.description': 'Still active in The Hague with 150+ members.',
-
-  // Pricing
-  'pricing.title': 'Membership & Pricing',
-  'pricing.popular': 'Popular',
-  'pricing.month': 'per month',
-  
-  // Plans
-  'pricing.plans.youth.name': 'Youth (up to 18 years)',
-  'pricing.plans.youth.price': '€27.50',
-  'pricing.plans.youth.f1': 'Unlimited training',
-  'pricing.plans.youth.f2': 'Access to all youth classes',
-  'pricing.plans.youth.f3': 'Exam opportunities',
-  'pricing.plans.youth.f4': 'Free judo suit up to 17 years',
-  
-  'pricing.plans.adults.name': 'Adults',
-  'pricing.plans.adults.price': '€27.50',
-  'pricing.plans.adults.f1': '2x per week training',
-  'pricing.plans.adults.f2': 'Access to all classes',
-  'pricing.plans.adults.f3': 'Exam opportunities',
-
-  'pricing.year': 'per year',
-  'pricing.plans.youth.yearly': '€330.-',
-  'pricing.plans.adults.yearly': '€330.-',
-
-  // Extra Info
-  'pricing.registrationFee': 'One-time Registration Fee',
-  'pricing.registrationFeeDescription': 'This fee is only paid once during your initial registration and is valid for your entire membership.',
-  'pricing.ooievaarspas': 'Discounts up to 100% on membership fees possible with the Ooievaarspas',
-  'pricing.startNow': 'Start Now',
-  'pricing.loading': 'Loading pricing...',
-  'pricing.error': 'Error loading pricing',
-  'pricing.noPrices': 'No pricing plans available',
-  'pricing.cta.title': 'Free Trial Lesson',
-  'pricing.cta.desc': 'Try a free trial lesson before signing up. No obligations!',
-  'pricing.cta.button': 'Book a Free Trial',
-  
-  // Exam
-  'exam.title': 'Exam Requirements',
-  "exam.participationTitle": "Participation is always encouraged",
-  "exam.participationText": "While specific minimum requirements apply for full belt grades, intermediate stripes can also be earned during exams. We therefore encourage everyone to participate.",
-  'exam.loading': 'Loading...',
-  'exam.error': 'Error loading exam requirements',
-  'exam.noGrades': 'No exam requirements available.',
-  'exam.download': 'Download Exam Document',
-  'exam.preview': 'Preview',
-  'exam.supplementary': 'Supplementary Documents',
-  'exam.minAge': 'Minimum age',
-
-  // Agenda
-  'agenda.badge': 'Year Calendar',
-  'agenda.title': 'Year Overview Agenda',
-  'agenda.loading': 'Loading...',
-  'agenda.error': 'Error loading agenda',
-  'agenda.noEvents': 'No events scheduled.',
-  'agenda.vacation': 'Vacation',
-  'agenda.holiday': 'Holiday',
-  'agenda.exam': 'Exam',
-  'agenda.competition': 'Competition',
-  'agenda.other': 'Event',
-  'agenda.allDay': 'All day',
-  'agenda.multiDay': 'Multi-day',
-  'agenda.registrationRequired': 'Registration required',
-  'agenda.registrationDeadline': 'Deadline',
-  'agenda.maxParticipants': 'Max participants',
-  'agenda.moreInfo': 'More information',
-  'agenda.location': 'Location',
-
-  // Media/Albums
-  'media.title': 'Photo Albums',
-  'media.description': 'View photos or videos from training sessions, events, competitions and special moments.',
-  'media.search': 'Search media albums...',
-  'media.loading': 'Loading albums...',
-  'media.error': 'Error loading albums',
-  'media.noAlbums': 'No albums available yet.',
-  'media.viewAlbum': 'View Album',
-  'media.close': 'Close',
-  'media.previous': 'Previous',
-  'media.next': 'Next',
-  'media.filterContent': 'All Albums',
-  'media.filterPhotos': 'Photos Only',
-  'media.filterVideos': 'With Videos',
-  'media.download': 'Download',
-  'media.downloadAll': 'Download all',
-  'media.downloadingAll': 'Downloading...',
-
-  // News month filter
-  'news.filterMonth': 'All months',
-  'news.month.1': 'January',
-  'news.month.2': 'February',
-  'news.month.3': 'March',
-  'news.month.4': 'April',
-  'news.month.5': 'May',
-  'news.month.6': 'June',
-  'news.month.7': 'July',
-  'news.month.8': 'August',
-  'news.month.9': 'September',
-  'news.month.10': 'October',
-  'news.month.11': 'November',
-  'news.month.12': 'December',
-
-  // Regels (Rules & Regulations)
-  'regels.title': 'Rules & Regulations',
-  'regels.description': 'Here you will find all regulations and house rules of Shi-Sei Sport. Please read these documents carefully to stay informed about the rules and regulations.',
-  'regels.infoTitle': 'Important Information',
-  'regels.infoText': 'All members and parents/guardians are expected to be aware of these regulations. Click on a document to read more details and use the download button to download the PDF file.',
-  'regels.download': 'Download PDF',
-  'regels.noDocuments': 'There are currently no documents available.',
-  'regels.error': 'Could not load documents. Please try again later.',
-
-  // VCP
-  'vcp.title': 'Confidential Counselor (VCP)',
-  'vcp.subtitle': 'Safe and enjoyable sports',
-  'vcp.contact': 'Contact',
-  'vcp.vcpName': 'VCP Name',
-  'vcp.email': 'Email',
-  'vcp.whatDoesVcpDo': 'What does the VCP do?',
-  'vcp.forWhom': 'Who is the VCP for?',
-  'vcp.whyContact': 'Why contact?',
-  'vcp.aboutVcp': 'About the VCP',
-  'vcp.preventivePolicy': 'Preventive Policy',
-  'vcp.crossingBehavior': 'What is transgressive behavior?',
-  'vcp.vcpTasks': 'VCP Tasks',
-
-  // Inschrijven (Enrollment)
-  'inschrijven.title': 'Enrollment',
-  'inschrijven.description': 'Become a member of Shi-Sei Sport. Choose how you would like to enroll below.',
-  'inschrijven.downloadOption': 'Download PDF',
-  'inschrijven.downloadOptionDesc': 'Download the enrollment form as PDF, fill it out and submit it to the club.',
-  'inschrijven.onlineOption': 'Enroll Online',
-  'inschrijven.onlineOptionDesc': 'Fill out the form directly online and receive an automatically filled PDF document.',
-  'inschrijven.formTitle': 'Online Enrollment Form',
-  'inschrijven.formDescription': 'Fill out the form below and receive an automatically filled PDF document.',
-  'inschrijven.infoTitle': 'How does it work?',
-  'inschrijven.step1': 'Download the form below',
-  'inschrijven.step2': 'Fill it out completely',
-  'inschrijven.step3': 'Submit it to the club or send it digitally',
-  'inschrijven.download': 'Download',
-  'inschrijven.noDocuments': 'There are currently no enrollment forms available.',
-  'inschrijven.error': 'Could not load forms. Please try again later.',
-  'inschrijven.ctaTitle': 'Need help with enrollment?',
-  'inschrijven.ctaText': 'Contact us if you have questions about the enrollment process or need help filling out the forms.',
-  'inschrijven.ctaButton': 'Contact Us',
-  'inschrijven.heroTagline': 'Start your judo journey today',
-  'inschrijven.heroSub': 'Join our club and discover the power of judo',
-
-  // Enrollment Form
-  'enrollment.form.personalInfo': 'Personal Information',
-  'enrollment.form.firstName': 'First Name',
-  'enrollment.form.middleName': 'Prefix',
-  'enrollment.form.lastName': 'Last Name',
-  'enrollment.form.dateOfBirth': 'Date of Birth',
-  'enrollment.form.email': 'Email',
-  'enrollment.form.phone': 'Phone Number',
-  'enrollment.form.address': 'Address',
-  'enrollment.form.street': 'Street Name',
-  'enrollment.form.houseNumber': 'House Number',
-  'enrollment.form.postalCode': 'Postal Code',
-  'enrollment.form.city': 'City',
-  'enrollment.form.emergencyContact': 'Emergency Contact',
-  'enrollment.form.contactName': 'Name',
-  'enrollment.form.contactPhone': 'Phone Number',
-  'enrollment.form.contactRelation': 'Relation',
-  'enrollment.form.relationPlaceholder': 'E.g. parent, partner, friend',
-  'enrollment.form.guardianName': 'Parent/Guardian Name',
-  'enrollment.form.guardianNamePlaceholder': 'Only fill in when registering a minor',
-  'enrollment.form.judoExperience': 'Judo Experience',
-  'enrollment.form.experience': 'Experience Level',
-  'enrollment.form.experienceBeginner': 'Beginner',
-  'enrollment.form.experienceSome': 'Some experience',
-  'enrollment.form.experienceAdvanced': 'Advanced',
-  'enrollment.form.judoGrade': 'Judo Grade',
-  'enrollment.form.judoGrade.none': 'None / Unknown',
-  'enrollment.form.judoGrade.6kyu': '6th kyu (White)',
-  'enrollment.form.judoGrade.5kyu': '5th kyu (Yellow)',
-  'enrollment.form.judoGrade.4kyu': '4th kyu (Orange)',
-  'enrollment.form.judoGrade.3kyu': '3rd kyu (Green)',
-  'enrollment.form.judoGrade.2kyu': '2nd kyu (Blue)',
-  'enrollment.form.judoGrade.1kyu': '1st kyu (Brown)',
-  'enrollment.form.judoGrade.dan': 'Dan (Black belt)',
-  'enrollment.form.preferredDays': 'Preferred Training Days',
-  'enrollment.form.day.monday': 'Monday',
-  'enrollment.form.day.wednesday': 'Wednesday',
-  'enrollment.form.day.thursday': 'Thursday',
-  'enrollment.form.day.saturday': 'Saturday',
-  'enrollment.form.medicalInfo': 'Medical Information',
-  'enrollment.form.medicalInfoPlaceholder': 'Allergies, medication, limitations...',
-  'enrollment.form.remarks': 'Remarks',
-  'enrollment.form.remarksPlaceholder': 'Any questions or comments...',
-  'enrollment.form.paymentMethod': 'Payment Method',
-  'enrollment.form.paymentMethodNote': 'Select your payment method. With the Ooievaarspas you receive a discount and do not need to fill out a direct debit authorization.',
-  'enrollment.form.ooievaarspas': 'I have an Ooievaarspas',
-  'enrollment.form.ooievaarspasNote': 'Check this if you have a valid Ooievaarspas.',
-  'enrollment.form.ooievaarspasNumber': 'Ooievaarspas number',
-  'enrollment.form.ooievaarspasNumberPlaceholder': 'Enter your Ooievaarspas number',
-  'enrollment.form.ooievaarspasCheckLocation': 'Your Ooievaarspas will be verified on location.',
-  'enrollment.form.machtiging': 'Direct Debit Authorization',
-  'enrollment.form.machtigingNote': 'For automatic payments, we need your bank account details.',
-  'enrollment.form.accountHolder': 'Account Holder Name',
-  'enrollment.form.iban': 'IBAN Account Number',
-  'enrollment.form.submit': 'Submit Enrollment',
-  'enrollment.form.submitting': 'Submitting...',
-  'enrollment.form.success': 'Enrollment Received!',
-  'enrollment.form.successText': 'Thank you for your enrollment! We have received your information and will contact you within 2 business days.',
-  'enrollment.form.pdfGenerated': 'A filled PDF form has been automatically generated and saved.',
-  'enrollment.form.error': 'Something went wrong while submitting. Please try again.',
-  'enrollment.form.captchaRequired': 'Please complete the CAPTCHA verification before submitting the form.',
-  'enrollment.form.agreeTermsPre': 'I agree to the',
-  'enrollment.form.agreeTermsLink': 'Terms & Conditions',
-  'enrollment.form.agreeTermsMid': 'and the',
-  'enrollment.form.agreePrivacyLink': 'Privacy Policy',
-  'enrollment.form.agreeTermsPost': 'of Shi-Sei Sport.',
-  'enrollment.form.signature': 'Signature',
-  'enrollment.form.signatureClear': 'Clear',
-  'enrollment.form.disclaimer': 'By submitting this form you agree to our privacy policy.',
-
-  // Enrollment Confirmation
-  'enrollment.confirm.title': 'Confirm Details',
-  'enrollment.confirm.description': 'Please review your details before submitting your enrollment.',
-  'enrollment.confirm.cancel': 'Go Back',
-  'enrollment.confirm.submit': 'Confirm & Submit',
-
-  // Trial
-  'trial.title': 'Book a Free Trial',
-  'trial.description': 'Try judo at Shi-Sei Sport. No obligations, just come and experience!',
-  'trial.benefits': 'What do you get with a trial lesson?',
-  'trial.benefit1': 'Free and without obligations',
-  'trial.benefit2': 'Meet our trainers',
-  'trial.benefit3': 'Experience our training style',
-  'trial.benefit4': 'Information about membership',
-  'trial.benefit5': 'Can start training immediately',
-  'trial.days': 'Training Days',
-  'trial.daysText': 'Monday, Wednesday, Thursday and Saturday',
-  'trial.scheduleLink': 'View full schedule',
-  'trial.bring': 'What to bring?',
-  'trial.bring1': 'Comfortable sportswear',
-  'trial.bring2': 'Water bottle',
-  'trial.bring3': 'Slippers',
-  'trial.bring4': 'Judo-gi will be loaned for the first time',
-  'trial.forWho': 'For whom?',
-  'trial.forWhoText': 'Everyone from 4 years old is welcome! Whether you are a beginner or already have experience, we have lessons for all levels.',
-  'trial.formTitle': 'Book your Trial',
-  'trial.age': 'Age',
-  'trial.experience': 'Experience',
-  'trial.experienceOptions.beginner': 'Beginner',
-  'trial.experienceOptions.some': 'Some experience',
-  'trial.experienceOptions.advanced': 'Advanced',
-  'trial.preferredDay': 'Preferred training day',
-  'trial.preferredDayOptions.maandag': 'Monday',
-  'trial.preferredDayOptions.woensdag': 'Wednesday',
-  'trial.preferredDayOptions.donderdag': 'Thursday',
-  'trial.preferredDayOptions.zaterdag': 'Saturday',
-  'trial.preferredDayOptions.none': 'No preference',
-  'trial.preferredDayAdultHint': 'Training for adults (18+) is only on Monday and Thursday.',
-  'trial.message': 'Comments (optional)',
-  'trial.messagePlaceholder': 'Do you have questions or specific wishes?',
-  'trial.button': 'Book Free Trial',
-  'trial.submitting': 'Submitting...',
-  'trial.error': 'An error occurred while submitting. Please try again later.',
-  'trial.captchaRequired': 'Please complete the CAPTCHA to continue.',
-  'trial.success': 'Thank you for your registration!',
-  'trial.successText': 'We have received your request and will contact you within 3 days to schedule a suitable time.',
-  'trial.disclaimer': 'By booking you agree to our privacy policy. We will contact you within 3 days.',
-  
-  // Footer
-  'footer.club': 'Shi-Sei Sport',
-  'footer.description': 'The oldest judo club in The Hague. Judo lessons since 1950.',
-  'footer.quickLinks': 'Quick Links',
-  'footer.contact': 'Contact',
-  'footer.contactText': 'For questions or more information, contact us via the contact page.',
-  'footer.jbn': 'Officially affiliated with the JBN.',
-  'footer.partners': 'Partners',
-  'footer.ooievaarspas': 'We accept the Ooievaarspas.',
-  'footer.copyright': 'All rights reserved.',
-  'footer.followUs': 'Follow us on',
-  'footer.privacy': 'Privacy Policy',
-  'footer.rules': 'House Rules',
-  'footer.terms': 'Terms & Conditions',
-
-  // Privacy Policy
-  'privacy.title': 'Privacy Policy',
-  'privacy.lastUpdated': 'Last updated',
-  'privacy.intro.title': 'Introduction',
-  'privacy.intro.text1': 'Shi-Sei Sport respects the privacy of all visitors to our website and members of our judo club. In this privacy policy we explain how we handle your personal data.',
-  'privacy.intro.text2': 'This privacy policy applies to the use of the website and services offered by Shi-Sei Sport.',
-  'privacy.collection.title': 'What data do we collect?',
-  'privacy.collection.intro': 'We collect different types of data for different purposes:',
-  'privacy.collection.item1': 'Contact information such as name, email address and phone number',
-  'privacy.collection.item2': 'Age and date of birth (for proper placement in training groups)',
-  'privacy.collection.item3': 'Experience and belt level (for proper placement in training groups)',
-  'privacy.collection.item4': 'Communication through our contact forms and email',
-  'privacy.collection.item5': 'Photos and visual material (for promotional purposes such as the website and social media)',
-  'privacy.usage.title': 'How do we use your data?',
-  'privacy.usage.intro': 'We use your personal data for the following purposes:',
-  'privacy.usage.item1': 'Processing your membership or trial lesson registration',
-  'privacy.usage.item2': 'Sending information about training sessions, events and exams',
-  'privacy.usage.item3': 'Answering questions and requests',
-  'privacy.usage.item4': 'Improving our services and website',
-  'privacy.protection.title': 'How do we protect your data?',
-  'privacy.protection.text1': 'We take the protection of your data seriously and take appropriate security measures to prevent misuse, loss and unauthorized access.',
-  'privacy.protection.text2': 'Your data is stored on secure servers and is only accessible to authorized personnel of Shi-Sei Sport.',
-  'privacy.rights.title': 'Your rights',
-  'privacy.rights.intro': 'You have the following rights regarding your personal data:',
-  'privacy.rights.item1': 'Right of access: you can request which data we process about you',
-  'privacy.rights.item2': 'Right to correction: you can ask to correct your data if it is incorrect',
-  'privacy.rights.item3': 'Right to erasure: you can request to delete your data',
-  'privacy.rights.item4': 'Right to object: you can object to the processing of your data',
-  'privacy.cookies.title': 'Cookies',
-  'privacy.cookies.text1': 'Our website only uses functional storage that is necessary for the proper functioning of the website. We do not use tracking cookies or analytics that track your browsing behavior.',
-  'privacy.cookies.intro': 'The following data is stored locally in your browser:',
-  'privacy.cookies.item1': 'Language preference (localStorage) — remembers whether you chose the Dutch or English version of the site.',
-  'privacy.cookies.item2': 'Dark mode (localStorage) — remembers whether you chose the light or dark theme.',
-  'privacy.contact.title': 'Contact',
-  'privacy.contact.text': 'Do you have questions about our privacy policy or would you like to exercise your rights? Please contact us:',
-  'privacy.contact.email': 'Email',
-
-  // Terms & Conditions
-  'terms.title': 'Terms & Conditions',
-  'terms.lastUpdated': 'Last updated',
-  'terms.intro.title': 'General Provisions',
-  'terms.intro.text': 'These terms and conditions apply to membership of Shi-Sei Sport. By becoming a member you accept these conditions.',
-
-  'terms.expect.title': 'What you can expect from Shi-Sei Sport',
-  'terms.expect.item1': 'Responsible sports in a safe environment.',
-  'terms.expect.item2': 'Your personal and address details entrusted to us will not be shared with third parties without consulting you.',
-  'terms.expect.item3': 'Lessons are taught by skilled, trained and/or enthusiastic coaches.',
-  'terms.expect.item4': 'All supervisors, volunteers and/or coaches hold a valid Certificate of Good Conduct (VOG).',
-  'terms.expect.item5': 'Each department has its own dedicated professional coach/supervisor.',
-  'terms.expect.item6': 'A professionally run sports club with adequate first aid resources.',
-  'terms.expect.item7': 'A great deal of enthusiasm from every leader.',
-
-  'terms.enrollment.title': 'Enrollment',
-  'terms.enrollment.item1': 'Please read the terms and conditions carefully before completing and submitting the enrollment form.',
-  'terms.enrollment.item2pre': 'Every member is required to take note of the',
-  'terms.enrollment.regelementLink': 'general regulations',
-  'terms.enrollment.huishoudelijkLink': 'house regulations',
-  'terms.enrollment.item2mid': 'and the',
-  'terms.enrollment.gedragscodeLink': 'code of conduct',
-
-  'terms.signature.title': 'Agreement',
-  'terms.signature.item1': 'By checking or signing the online enrollment form, the undersigned declares to agree with the membership conditions of Shi-Sei Sport, of which the General Terms and the dojo rules are an inseparable part, and understands what is stated therein.',
-  'terms.signature.item2': 'The General Terms are visible on the website www.shi-sei.nl.',
-  'terms.signature.item3': 'The rights arising from the membership are strictly personal.',
-  'terms.signature.item4': 'Persons under 18 years of age may only be registered by their parent(s) or legal guardian(s).',
-  'terms.signature.item5': 'By signing the direct debit authorization, you grant Shi-Sei Sport permission to automatically debit the amounts due for advance payment of the membership fee and other amounts owed to Shi-Sei Sport from your specified bank account until further notice.',
-
-  'terms.membership.title': 'Membership',
-  'terms.membership.item1': 'Membership commences upon receipt of registration and payment of the membership fee.',
-  'terms.membership.item2': 'Persons under 18 years of age may only be registered by their parent(s) or legal guardian(s).',
-  'terms.membership.item3pre': 'Members must adhere to the',
-  'terms.membership.item3mid': 'and',
-  'terms.membership.item4': 'Membership is personal and non-transferable.',
-
-  'terms.link.huisregels': 'house rules',
-  'terms.link.gedragscode': 'code of conduct',
-
-  'terms.rates.title': 'Pricing',
-  'terms.rates.item1': 'Rates may be adjusted by Shi-Sei Sport if necessary.',
-  'terms.rates.item2': 'Changes to membership rates will be communicated by email 1 month before they take effect.',
-  'terms.rates.item3': 'In the event of a price increase, the member has the right to terminate the agreement within 4 weeks of the announcement.',
-  'terms.rates.item4': 'Any prepaid subscription fees will be refunded.',
-
-  'terms.payment.title': 'Payment',
-  'terms.payment.item1': 'The member or their representative is responsible for ensuring smooth payment of membership contributions.',
-  'terms.payment.item2': 'Membership contributions are automatically collected quarterly around the first of the new month.',
-  'terms.payment.item3': 'Contributions are paid quarterly in advance.',
-  'terms.payment.item4': 'In case of late payment, access to classes may be suspended.',
-  'terms.payment.item5': 'Refund of membership fees is not possible, except in exceptional cases.',
-
-  'terms.behavior.title': 'Behavior and House Rules',
-  'terms.behavior.item1': 'Members must treat other members, trainers and equipment with respect.',
-  'terms.behavior.item2pre': 'Members must adhere to the',
-  'terms.behavior.item2mid': 'and',
-  'terms.behavior.item3': 'Participation in training under the influence of alcohol or drugs is not permitted.',
-  'terms.behavior.item4': 'In case of serious violations, the board may decide to suspend or terminate membership.',
-
-  'terms.exams.title': 'Exams and Competitions',
-  'terms.exams.item1': 'The costs for the special competition group, competitions and exams are not included in the lesson fees.',
-  'terms.exams.item2': 'In the event of absence of a member at an exam or competition, no refund of the registration fee will be given.',
-
-  'terms.liability.title': 'Own Risk',
-  'terms.liability.text1': 'Shi-Sei Sport is not liable for injury, loss or damage that occurs during or as a result of sports activities, unless this is the result of gross negligence or intent by the association.',
-  'terms.liability.text2': 'Members are strongly advised to take out their own sports injury or accident insurance.',
-
-  'terms.photos.title': 'Photos and Media',
-  'terms.photos.item1': 'The member and other visitors agree to the use of their image in print, photo, video, etc. for the website and other promotional purposes without claiming compensation.',
-  'terms.photos.item2': 'If you object to this, you can indicate this in writing on the enrollment form and communicate any changes by email at any time.',
-
-  'terms.changes.title': 'Changes',
-  'terms.changes.text': 'Shi-Sei Sport reserves the right to change these terms and conditions. Members will be notified of this via the website or by email.',
-
-  'terms.cancellation.title': 'Cancellation',
-  'terms.cancellation.item1': 'Membership can be terminated with one month\'s notice.',
-  'terms.cancellation.item2': 'Cancellation must be done in writing or by email.',
-  'terms.cancellation.item3': 'If cancelled within the contract period, membership fees remain due until the end of that period.',
-
-  'terms.contact.title': 'Questions',
-  'terms.contact.text': 'Do you have questions about these terms and conditions? Please contact us:',
-  'terms.contact.email': 'Email',
-
-  // Common
-  'common.loading': 'Loading...',
-  'common.required': '*',
-  'common.select': 'Select',
-  'common.optional': '(optional)',
-  'common.invalidEmail': 'Please enter a valid email address, e.g. name@example.com',
-  'common.invalidPhone': 'Please enter a valid phone number, e.g. +31-612345678',
-  'common.invalidIban': 'Please enter a valid IBAN, e.g. NL91 ABNA 0417 1643 00',
-  'common.invalidPostalCode': 'Please enter a valid postal code, e.g. 1234 AB',
-  'common.invalidAge': 'Please enter a valid age (4–100)',
-  'common.ibanCountryHint': 'The form automatically adjusts to the country code.',
-
-  // Placeholders
-  'placeholder.name': 'Your full name',
-  'placeholder.firstName': 'First name',
-  'placeholder.middleName': 'Prefix (optional)',
-  'placeholder.lastName': 'Last name',
-  'placeholder.email': 'name@example.com',
-  'placeholder.phone': '06-12345678',
-  'placeholder.street': 'Street name',
-  'placeholder.houseNumber': 'No.',
-  'placeholder.city': 'City',
-  'placeholder.accountHolder': 'Account holder name',
-  'placeholder.age': 'e.g. 8',
-  'placeholder.message': 'Type your message here...',
-
-  // 404
-  '404.title': 'Page not found',
-  '404.description': 'The page you are looking for does not exist or has been moved.',
-  '404.home': 'Go to Homepage',
-  '404.back': 'Back',
 };
