+++ frontend/src/i18n/en.ts
@@ -1,0 +1,592 @@
+const en: Record<string, string> = {
+  // Navigation
+  'nav.home': 'Home',
+  'nav.about': 'About Us',
+  'nav.team': 'The Team',
+  'nav.history': 'History',
+  'nav.locations': 'Locations',
+  'nav.rules': 'Rules & Regulations',
+  'nav.info': 'Information',
+  'nav.schedule': 'Schedule',
+  'nav.pricing': 'Pricing',
+  'nav.exam': 'Exam Requirements',
+  'nav.enrollment': 'Enrollment',
+  'nav.agenda': 'Agenda',
+  'nav.news': 'News',
+  'nav.media': 'Media',
+  'nav.contact': 'Contact',
+  'nav.trial': 'Free Trial',
+
+  // Hero
+  'hero.title': 'Judo',
+  'hero.subtitle': 'SHI-SEI SPORT',
+  'hero.description': 'The oldest judo club in The Hague',
+  'hero.text': 'We teach judo on Monday, Wednesday, Thursday and Saturday. Experience the power of discipline, respect and self-confidence in our dojo.',
+  'hero.button': 'Try a Trial Lesson',
+  'hero.scheduleButton': 'View the Schedule',
+
+  // News
+  'news.latest': 'Latest News',
+  'news.all': 'All news',
+  'news.search': 'Search news...',
+  'news.readMore': 'Read more',
+  'news.clickToRead': 'Click to read the full article...',
+  'news.loading': 'Loading...',
+  'news.notFound': 'News not found',
+  'news.back': 'Back to news',
+  'news.backAll': 'Back to all news',
+  'news.noNews': 'No news available yet.',
+  'news.title': 'Latest News',
+  'news.description': 'Stay up to date with the latest news and events at Shi-Sei Sport.',
+  'news.share': 'Share this article',
+  'news.shareCopied': 'Link copied!',
+  'news.noContent': 'This article has no content.',
+
+  // Schedule
+  'schedule.title': 'Lessons & Schedule',
+  'schedule.description': 'We teach judo on Monday, Wednesday, Thursday and Saturday',
+  'schedule.loading': 'Loading schedule...',
+  'schedule.error': 'Could not load schedule. Please try again later.',
+  'schedule.retry': 'Try again',
+  'schedule.noClasses': 'No classes scheduled yet.',
+
+  // Locations
+  'locations.title': 'Our Training Locations',
+  'locations.address': 'Address',
+  'locations.noMap': 'No map available',
+  'locations.noLocations': 'No locations available',
+  'locations.loading': 'Loading locations...',
+  'locations.error': 'Could not load locations. Please try again later.',
+  'locations.questionsTitle': 'Questions about our locations?',
+  'locations.postalAddress': 'Postal Address',
+  'locations.phone': 'Phone',
+  'locations.email': 'Email',
+  'locations.hours': 'Opening Hours',
+  'locations.hoursText': 'Depends on class schedule',
+  'locations.contactTitle': 'Questions or Need Contact?',
+  'locations.contactDescription': 'Contact us for more information about our training locations, opening hours, or other questions.',
+  'locations.contactButton': 'Go to Contact Page',
+  'locations.openMaps': 'Open in Google Maps',
+
+  // Contact
+  'contact.title': 'Contact Us',
+  'contact.description': 'Have questions? We are here for you! Fill out the form or contact us directly.',
+  'contact.info': 'Contact Information',
+  'contact.address': 'Postal Address',
+  'contact.phone': 'Phone',
+  'contact.email': 'Email',
+  'contact.hours': 'Training Hours',
+  'contact.hoursText': 'Monday, Wednesday, Thursday and Saturday',
+  'contact.fullScheduleText': 'View the full schedule on the',
+  'contact.scheduleLink': 'schedule page',
+  'contact.formTitle': 'Send a Message',
+  'contact.name': 'Name',
+  'contact.subject': 'Subject',
+  'contact.message': 'Message',
+  'contact.send': 'Send Message',
+  'contact.sending': 'Sending...',
+  'contact.error': 'Something went wrong while sending. Please try again.',
+  'contact.captchaRequired': 'Please complete the CAPTCHA verification before submitting the form.',
+  'common.captchaLabel': "I'm not a robot, but a Judoka",
+  'contact.success': 'Message Received!',
+  'contact.successText': 'Thank you for your message! We will contact you as soon as possible.',
+  'contact.subjectOptions.proefles': 'Free Trial',
+  'contact.subjectOptions.inschrijving': 'Registration',
+  'contact.subjectOptions.vraag': 'General Question',
+  'contact.subjectOptions.anders': 'Other',
+  'contact.subjectSelect': 'Select a subject',
+
+  // Team
+  'team.title': 'Our Team',
+  'team.description': 'Meet our experienced and passionate trainers who are committed to your development in judo.',
+  'team.cta': 'Interested in Training?',
+  'team.ctaText': 'Come by for a free trial lesson and meet our team!',
+  'team.button': 'Book a Free Trial',
+  'team.error': 'An error occurred while loading the team.',
+  'team.noInstructors': 'No instructors found.',
+  'team.qualifications': 'Certifications & Achievements:',
+
+  'history.title': 'Our Rich History',
+
+  // Section 1: The Foundation
+  'history.section1.title': 'The Foundation (1950)',
+  'history.section1.p1': 'In November 1950, JUDO ASSOCIATION SHI-SEI was founded. A group of enthusiastic judokas from Johan v.d. Bruggen in the Zoutmanstraat wanted their own club. A place where they could train more often and at a lower cost.',
+  'history.section1.p2': 'Led by Jacques Brakel, Hans van Diggelen and several others, Judo Association SHI-SEI was born on November 5, 1950, registered with the Chamber of Commerce and affiliated with the Dutch Amateur Judo Association.',
+  'history.section1.p3': 'Initially, training took place in a gym on Daal en Bergselaan, but the club soon moved to a permanent dojo at Jan Hendrikstraat 9. Here, members could train to their heart\'s content, and NAJA national trainer Tokio Hirano, 6th dan, ensured the improvement of techniques. Training was now possible seven days a week, all day if necessary. It didn\'t take long for the training to yield results.',
+  'history.section1.tokioHiranoLink': 'https://judoinfo.com/hirano/#:~:text=Tokio%20Hirano:%20The%20Man%20Who%20Revolutionized%20Judo%20%7C%20Judo%20Info.%20Judo%20Info.%20Online%20Dojo.',
+  'history.section1.p4': 'Wil Wagner was the club\'s first black belt holder. At a judo summer school in Switzerland, he received the coveted belt from grandmaster Hanno Rhi. Shortly after, Jan v.d. Toorn also achieved the 1st dan grade (black belt) as the second member. Both of them provided the technical leadership on the mat during the early period, supplemented by brown belts who acted as instructors on various evenings.',
+  'history.section1.p5': 'The organizational leadership under chairman Jacques Brakel ensured the club ran excellently in that regard as well. Black and brown belts were enrolled in Instructor and referee courses, thus providing the necessary "know-how".',
+
+  // Highlight Box: Growth in the Sixties
+  'history.growth.title': 'Growth in the Sixties',
+  'history.growth.description': 'In the sixties, SHI-SEI was a leading club in The Hague and a formidable opponent on the mat. The membership grew steadily larger and the staff also expanded continuously.',
+  'history.growth.didYouKnow': 'Did you know that SHI-SEI became the mother club of many Hague judo clubs such as Lu-Gia-Jen, Guan-Tsui, Huan-Pi, De Doorkruiers, Guan Lin, Kagawa, and Sumeju? Wil Wagner was the founder of most of these clubs.',
+
+  // Section 2: Successful Judokas
+  'history.section2.title': 'Successful Judokas',
+  'history.section2.p1': 'The club has produced hundreds of black belt holders, including a number of particularly good competition judokas such as Wil Tap, Jan Hendriks, Jaap Zoutendijk, Jan Dotinga, Kees Toonen, Koos Tap, Koos Letterie, Wim Lut, Giel Groos, Beb de Waal, Nico Schoemaker, Koos v.d. Mark, Wil Wagner, Jan v.d. Toorn, Ed Beekink, Wim de Rooy, Barry Oudshoorn, Henk Ammerlaan, John Lut and many others.',
+
+  // Section 3: Resurrection & New Locations
+  'history.section3.title': 'Resurrection & New Locations',
+  'history.section3.p1': 'In 1970, the association almost ceased to exist when a number of members dismissed the entire board at an extraordinary general meeting. This turned out to be a disastrous failure.',
+  'history.section3.p2': 'However, in 1972 a new board was formed under the inspiring leadership of Wim Lut, and the club was given new life, particularly through branches in Westland. This, however, resulted in the accommodation becoming far too small and no longer meeting the requirements of the time.',
+  'history.section3.p3': 'Finally, in 1981, through aldermen Piet Vink and Adrie Duivestein, we obtained a spacious accommodation in the old school at Nieuwe Schoolstraat 22-B.',
+  'history.section3.p4': 'Here we have a spacious training hall, two spacious shower facilities, office, canteen, small training hall and plenty of storage space. In this space we now have not only judo but also, since 1990, a taekwondo department with various home-grown 1st and 2nd dans. A darts department also provides the necessary conviviality.',
+
+  // Section 4: Relocations & Recent History
+  'history.section4.title': 'Relocations & Recent History',
+  'history.section4.p1': 'Unfortunately, our beautiful own location came to an end in 2006 when the municipality decided to terminate our lease contract to build three luxury apartments on the site of our beautiful location. Unfortunately, we could not find our own new location that we could afford.',
+  'history.section4.p2': 'We decided to use a community center in Fluitenbergstraat in The Hague and a gym at Beresteinlaan in The Hague. In 2011, the Morgenstond meeting center became our new location and the gym at Beresteinlaan was once again necessarily exchanged for a nicer gym at Pachtersdreef.',
+  'history.section4.p3': 'From 2011 until the corona crisis in 2020, we taught judo and taekwondo 6 days a week. Our membership had grown steadily to approximately 140 judo members and 40 taekwondo members.',
+  'history.section4.p4': 'From March 16, 2020, we were forced to close for a short period due to the corona virus infection. But now we are back to training at full speed.',
+
+  // Timeline & Stats
+  'history.timeline.title': 'Timeline',
+  'history.stats.title': 'Current Status',
+  'history.stats.members': 'Members',
+  'history.stats.region': 'Location',
+  'history.stats.city': 'The Hague',
+
+  'history.milestones.1950.title': 'Foundation',
+  'history.milestones.1950.description': 'Start at Daal en Bergselaan.',
+  'history.milestones.1960.title': 'Golden Years',
+  'history.milestones.1960.description': 'Mother club of many Hague associations.',
+  'history.milestones.1980.title': 'Nieuwe Schoolstraat',
+  'history.milestones.1980.description': 'Relocation to the iconic dojo (1981).',
+  'history.milestones.2011.title': 'Morgenstond',
+  'history.milestones.2011.description': 'New location at Pachtersdreef.',
+  'history.milestones.2025.title': 'Present',
+  'history.milestones.2025.description': 'Still active in The Hague with 150+ members.',
+
+  // Pricing
+  'pricing.title': 'Membership & Pricing',
+  'pricing.popular': 'Popular',
+  'pricing.month': 'per month',
+
+  // Plans
+  'pricing.plans.youth.name': 'Youth (up to 18 years)',
+  'pricing.plans.youth.price': '\u20AC27.50',
+  'pricing.plans.youth.f1': 'Unlimited training',
+  'pricing.plans.youth.f2': 'Access to all youth classes',
+  'pricing.plans.youth.f3': 'Exam opportunities',
+  'pricing.plans.youth.f4': 'Free judo suit up to 17 years',
+
+  'pricing.plans.adults.name': 'Adults',
+  'pricing.plans.adults.price': '\u20AC27.50',
+  'pricing.plans.adults.f1': '2x per week training',
+  'pricing.plans.adults.f2': 'Access to all classes',
+  'pricing.plans.adults.f3': 'Exam opportunities',
+
+  'pricing.year': 'per year',
+  'pricing.plans.youth.yearly': '\u20AC330.-',
+  'pricing.plans.adults.yearly': '\u20AC330.-',
+
+  // Extra Info
+  'pricing.registrationFee': 'One-time Registration Fee',
+  'pricing.registrationFeeDescription': 'This fee is only paid once during your initial registration and is valid for your entire membership.',
+  'pricing.ooievaarspas': 'Discounts up to 100% on membership fees possible with the Ooievaarspas',
+  'pricing.startNow': 'Start Now',
+  'pricing.loading': 'Loading pricing...',
+  'pricing.error': 'Error loading pricing',
+  'pricing.noPrices': 'No pricing plans available',
+  'pricing.cta.title': 'Free Trial Lesson',
+  'pricing.cta.desc': 'Try a free trial lesson before signing up. No obligations!',
+  'pricing.cta.button': 'Book a Free Trial',
+
+  // Exam
+  'exam.title': 'Exam Requirements',
+  "exam.participationTitle": "Participation is always encouraged",
+  "exam.participationText": "While specific minimum requirements apply for full belt grades, intermediate stripes can also be earned during exams. We therefore encourage everyone to participate.",
+  'exam.loading': 'Loading...',
+  'exam.error': 'Error loading exam requirements',
+  'exam.noGrades': 'No exam requirements available.',
+  'exam.download': 'Download Exam Document',
+  'exam.preview': 'Preview',
+  'exam.supplementary': 'Supplementary Documents',
+  'exam.minAge': 'Minimum age',
+
+  // Agenda
+  'agenda.badge': 'Year Calendar',
+  'agenda.title': 'Year Overview Agenda',
+  'agenda.loading': 'Loading...',
+  'agenda.error': 'Error loading agenda',
+  'agenda.noEvents': 'No events scheduled.',
+  'agenda.vacation': 'Vacation',
+  'agenda.holiday': 'Holiday',
+  'agenda.exam': 'Exam',
+  'agenda.competition': 'Competition',
+  'agenda.other': 'Event',
+  'agenda.allDay': 'All day',
+  'agenda.multiDay': 'Multi-day',
+  'agenda.registrationRequired': 'Registration required',
+  'agenda.registrationDeadline': 'Deadline',
+  'agenda.maxParticipants': 'Max participants',
+  'agenda.moreInfo': 'More information',
+  'agenda.location': 'Location',
+
+  // Media/Albums
+  'media.title': 'Photo Albums',
+  'media.description': 'View photos or videos from training sessions, events, competitions and special moments.',
+  'media.search': 'Search media albums...',
+  'media.loading': 'Loading albums...',
+  'media.error': 'Error loading albums',
+  'media.noAlbums': 'No albums available yet.',
+  'media.viewAlbum': 'View Album',
+  'media.close': 'Close',
+  'media.previous': 'Previous',
+  'media.next': 'Next',
+  'media.filterContent': 'All Albums',
+  'media.filterPhotos': 'Photos Only',
+  'media.filterVideos': 'With Videos',
+  'media.download': 'Download',
+  'media.downloadAll': 'Download all',
+  'media.downloadingAll': 'Downloading...',
+
+  // News month filter
+  'news.filterMonth': 'All months',
+  'news.month.1': 'January',
+  'news.month.2': 'February',
+  'news.month.3': 'March',
+  'news.month.4': 'April',
+  'news.month.5': 'May',
+  'news.month.6': 'June',
+  'news.month.7': 'July',
+  'news.month.8': 'August',
+  'news.month.9': 'September',
+  'news.month.10': 'October',
+  'news.month.11': 'November',
+  'news.month.12': 'December',
+
+  // Regels (Rules & Regulations)
+  'regels.title': 'Rules & Regulations',
+  'regels.description': 'Here you will find all regulations and house rules of Shi-Sei Sport. Please read these documents carefully to stay informed about the rules and regulations.',
+  'regels.infoTitle': 'Important Information',
+  'regels.infoText': 'All members and parents/guardians are expected to be aware of these regulations. Click on a document to read more details and use the download button to download the PDF file.',
+  'regels.download': 'Download PDF',
+  'regels.noDocuments': 'There are currently no documents available.',
+  'regels.error': 'Could not load documents. Please try again later.',
+
+  // VCP
+  'vcp.title': 'Confidential Counselor (VCP)',
+  'vcp.subtitle': 'Safe and enjoyable sports',
+  'vcp.contact': 'Contact',
+  'vcp.vcpName': 'VCP Name',
+  'vcp.email': 'Email',
+  'vcp.whatDoesVcpDo': 'What does the VCP do?',
+  'vcp.forWhom': 'Who is the VCP for?',
+  'vcp.whyContact': 'Why contact?',
+  'vcp.aboutVcp': 'About the VCP',
+  'vcp.preventivePolicy': 'Preventive Policy',
+  'vcp.crossingBehavior': 'What is transgressive behavior?',
+  'vcp.vcpTasks': 'VCP Tasks',
+
+  // Inschrijven (Enrollment)
+  'inschrijven.title': 'Enrollment',
+  'inschrijven.description': 'Become a member of Shi-Sei Sport. Choose how you would like to enroll below.',
+  'inschrijven.downloadOption': 'Download PDF',
+  'inschrijven.downloadOptionDesc': 'Download the enrollment form as PDF, fill it out and submit it to the club.',
+  'inschrijven.onlineOption': 'Enroll Online',
+  'inschrijven.onlineOptionDesc': 'Fill out the form directly online and receive an automatically filled PDF document.',
+  'inschrijven.formTitle': 'Online Enrollment Form',
+  'inschrijven.formDescription': 'Fill out the form below and receive an automatically filled PDF document.',
+  'inschrijven.infoTitle': 'How does it work?',
+  'inschrijven.step1': 'Download the form below',
+  'inschrijven.step2': 'Fill it out completely',
+  'inschrijven.step3': 'Submit it to the club or send it digitally',
+  'inschrijven.download': 'Download',
+  'inschrijven.noDocuments': 'There are currently no enrollment forms available.',
+  'inschrijven.error': 'Could not load forms. Please try again later.',
+  'inschrijven.ctaTitle': 'Need help with enrollment?',
+  'inschrijven.ctaText': 'Contact us if you have questions about the enrollment process or need help filling out the forms.',
+  'inschrijven.ctaButton': 'Contact Us',
+  'inschrijven.heroTagline': 'Start your judo journey today',
+  'inschrijven.heroSub': 'Join our club and discover the power of judo',
+
+  // Enrollment Form
+  'enrollment.form.personalInfo': 'Personal Information',
+  'enrollment.form.firstName': 'First Name',
+  'enrollment.form.middleName': 'Prefix',
+  'enrollment.form.lastName': 'Last Name',
+  'enrollment.form.dateOfBirth': 'Date of Birth',
+  'enrollment.form.email': 'Email',
+  'enrollment.form.phone': 'Phone Number',
+  'enrollment.form.address': 'Address',
+  'enrollment.form.street': 'Street Name',
+  'enrollment.form.houseNumber': 'House Number',
+  'enrollment.form.postalCode': 'Postal Code',
+  'enrollment.form.city': 'City',
+  'enrollment.form.emergencyContact': 'Emergency Contact',
+  'enrollment.form.contactName': 'Name',
+  'enrollment.form.contactPhone': 'Phone Number',
+  'enrollment.form.contactRelation': 'Relation',
+  'enrollment.form.relationPlaceholder': 'E.g. parent, partner, friend',
+  'enrollment.form.guardianName': 'Parent/Guardian Name',
+  'enrollment.form.guardianNamePlaceholder': 'Only fill in when registering a minor',
+  'enrollment.form.judoExperience': 'Judo Experience',
+  'enrollment.form.experience': 'Experience Level',
+  'enrollment.form.experienceBeginner': 'Beginner',
+  'enrollment.form.experienceSome': 'Some experience',
+  'enrollment.form.experienceAdvanced': 'Advanced',
+  'enrollment.form.judoGrade': 'Judo Grade',
+  'enrollment.form.judoGrade.none': 'None / Unknown',
+  'enrollment.form.judoGrade.6kyu': '6th kyu (White)',
+  'enrollment.form.judoGrade.5kyu': '5th kyu (Yellow)',
+  'enrollment.form.judoGrade.4kyu': '4th kyu (Orange)',
+  'enrollment.form.judoGrade.3kyu': '3rd kyu (Green)',
+  'enrollment.form.judoGrade.2kyu': '2nd kyu (Blue)',
+  'enrollment.form.judoGrade.1kyu': '1st kyu (Brown)',
+  'enrollment.form.judoGrade.dan': 'Dan (Black belt)',
+  'enrollment.form.preferredDays': 'Preferred Training Days',
+  'enrollment.form.day.monday': 'Monday',
+  'enrollment.form.day.wednesday': 'Wednesday',
+  'enrollment.form.day.thursday': 'Thursday',
+  'enrollment.form.day.saturday': 'Saturday',
+  'enrollment.form.medicalInfo': 'Medical Information',
+  'enrollment.form.medicalInfoPlaceholder': 'Allergies, medication, limitations...',
+  'enrollment.form.remarks': 'Remarks',
+  'enrollment.form.remarksPlaceholder': 'Any questions or comments...',
+  'enrollment.form.paymentMethod': 'Payment Method',
+  'enrollment.form.paymentMethodNote': 'Select your payment method. With the Ooievaarspas you receive a discount and do not need to fill out a direct debit authorization.',
+  'enrollment.form.ooievaarspas': 'I have an Ooievaarspas',
+  'enrollment.form.ooievaarspasNote': 'Check this if you have a valid Ooievaarspas.',
+  'enrollment.form.ooievaarspasNumber': 'Ooievaarspas number',
+  'enrollment.form.ooievaarspasNumberPlaceholder': 'Enter your Ooievaarspas number',
+  'enrollment.form.ooievaarspasCheckLocation': 'Your Ooievaarspas will be verified on location.',
+  'enrollment.form.machtiging': 'Direct Debit Authorization',
+  'enrollment.form.machtigingNote': 'For automatic payments, we need your bank account details.',
+  'enrollment.form.accountHolder': 'Account Holder Name',
+  'enrollment.form.iban': 'IBAN Account Number',
+  'enrollment.form.submit': 'Submit Enrollment',
+  'enrollment.form.submitting': 'Submitting...',
+  'enrollment.form.success': 'Enrollment Received!',
+  'enrollment.form.successText': 'Thank you for your enrollment! We have received your information and will contact you within 2 business days.',
+  'enrollment.form.pdfGenerated': 'A filled PDF form has been automatically generated and saved.',
+  'enrollment.form.error': 'Something went wrong while submitting. Please try again.',
+  'enrollment.form.captchaRequired': 'Please complete the CAPTCHA verification before submitting the form.',
+  'enrollment.form.agreeTermsPre': 'I agree to the',
+  'enrollment.form.agreeTermsLink': 'Terms & Conditions',
+  'enrollment.form.agreeTermsMid': 'and the',
+  'enrollment.form.agreePrivacyLink': 'Privacy Policy',
+  'enrollment.form.agreeTermsPost': 'of Shi-Sei Sport.',
+  'enrollment.form.signature': 'Signature',
+  'enrollment.form.signatureClear': 'Clear',
+  'enrollment.form.disclaimer': 'By submitting this form you agree to our privacy policy.',
+
+  // Enrollment Confirmation
+  'enrollment.confirm.title': 'Confirm Details',
+  'enrollment.confirm.description': 'Please review your details before submitting your enrollment.',
+  'enrollment.confirm.cancel': 'Go Back',
+  'enrollment.confirm.submit': 'Confirm & Submit',
+
+  // Trial
+  'trial.title': 'Book a Free Trial',
+  'trial.description': 'Try judo at Shi-Sei Sport. No obligations, just come and experience!',
+  'trial.benefits': 'What do you get with a trial lesson?',
+  'trial.benefit1': 'Free and without obligations',
+  'trial.benefit2': 'Meet our trainers',
+  'trial.benefit3': 'Experience our training style',
+  'trial.benefit4': 'Information about membership',
+  'trial.benefit5': 'Can start training immediately',
+  'trial.days': 'Training Days',
+  'trial.daysText': 'Monday, Wednesday, Thursday and Saturday',
+  'trial.scheduleLink': 'View full schedule',
+  'trial.bring': 'What to bring?',
+  'trial.bring1': 'Comfortable sportswear',
+  'trial.bring2': 'Water bottle',
+  'trial.bring3': 'Slippers',
+  'trial.bring4': 'Judo-gi will be loaned for the first time',
+  'trial.forWho': 'For whom?',
+  'trial.forWhoText': 'Everyone from 4 years old is welcome! Whether you are a beginner or already have experience, we have lessons for all levels.',
+  'trial.formTitle': 'Book your Trial',
+  'trial.age': 'Age',
+  'trial.experience': 'Experience',
+  'trial.experienceOptions.beginner': 'Beginner',
+  'trial.experienceOptions.some': 'Some experience',
+  'trial.experienceOptions.advanced': 'Advanced',
+  'trial.preferredDay': 'Preferred training day',
+  'trial.preferredDayOptions.maandag': 'Monday',
+  'trial.preferredDayOptions.woensdag': 'Wednesday',
+  'trial.preferredDayOptions.donderdag': 'Thursday',
+  'trial.preferredDayOptions.zaterdag': 'Saturday',
+  'trial.preferredDayOptions.none': 'No preference',
+  'trial.preferredDayAdultHint': 'Training for adults (18+) is only on Monday and Thursday.',
+  'trial.message': 'Comments (optional)',
+  'trial.messagePlaceholder': 'Do you have questions or specific wishes?',
+  'trial.button': 'Book Free Trial',
+  'trial.submitting': 'Submitting...',
+  'trial.error': 'An error occurred while submitting. Please try again later.',
+  'trial.captchaRequired': 'Please complete the CAPTCHA to continue.',
+  'trial.success': 'Thank you for your registration!',
+  'trial.successText': 'We have received your request and will contact you within 3 days to schedule a suitable time.',
+  'trial.disclaimer': 'By booking you agree to our privacy policy. We will contact you within 3 days.',
+
+  // Footer
+  'footer.club': 'Shi-Sei Sport',
+  'footer.description': 'The oldest judo club in The Hague. Judo lessons since 1950.',
+  'footer.quickLinks': 'Quick Links',
+  'footer.contact': 'Contact',
+  'footer.contactText': 'For questions or more information, contact us via the contact page.',
+  'footer.jbn': 'Officially affiliated with the JBN.',
+  'footer.partners': 'Partners',
+  'footer.ooievaarspas': 'We accept the Ooievaarspas.',
+  'footer.copyright': 'All rights reserved.',
+  'footer.followUs': 'Follow us on',
+  'footer.privacy': 'Privacy Policy',
+  'footer.rules': 'House Rules',
+  'footer.terms': 'Terms & Conditions',
+
+  // Privacy Policy
+  'privacy.title': 'Privacy Policy',
+  'privacy.lastUpdated': 'Last updated',
+  'privacy.intro.title': 'Introduction',
+  'privacy.intro.text1': 'Shi-Sei Sport respects the privacy of all visitors to our website and members of our judo club. In this privacy policy we explain how we handle your personal data.',
+  'privacy.intro.text2': 'This privacy policy applies to the use of the website and services offered by Shi-Sei Sport.',
+  'privacy.collection.title': 'What data do we collect?',
+  'privacy.collection.intro': 'We collect different types of data for different purposes:',
+  'privacy.collection.item1': 'Contact information such as name, email address and phone number',
+  'privacy.collection.item2': 'Age and date of birth (for proper placement in training groups)',
+  'privacy.collection.item3': 'Experience and belt level (for proper placement in training groups)',
+  'privacy.collection.item4': 'Communication through our contact forms and email',
+  'privacy.collection.item5': 'Photos and visual material (for promotional purposes such as the website and social media)',
+  'privacy.usage.title': 'How do we use your data?',
+  'privacy.usage.intro': 'We use your personal data for the following purposes:',
+  'privacy.usage.item1': 'Processing your membership or trial lesson registration',
+  'privacy.usage.item2': 'Sending information about training sessions, events and exams',
+  'privacy.usage.item3': 'Answering questions and requests',
+  'privacy.usage.item4': 'Improving our services and website',
+  'privacy.protection.title': 'How do we protect your data?',
+  'privacy.protection.text1': 'We take the protection of your data seriously and take appropriate security measures to prevent misuse, loss and unauthorized access.',
+  'privacy.protection.text2': 'Your data is stored on secure servers and is only accessible to authorized personnel of Shi-Sei Sport.',
+  'privacy.rights.title': 'Your rights',
+  'privacy.rights.intro': 'You have the following rights regarding your personal data:',
+  'privacy.rights.item1': 'Right of access: you can request which data we process about you',
+  'privacy.rights.item2': 'Right to correction: you can ask to correct your data if it is incorrect',
+  'privacy.rights.item3': 'Right to erasure: you can request to delete your data',
+  'privacy.rights.item4': 'Right to object: you can object to the processing of your data',
+  'privacy.cookies.title': 'Cookies',
+  'privacy.cookies.text1': 'Our website only uses functional storage that is necessary for the proper functioning of the website. We do not use tracking cookies or analytics that track your browsing behavior.',
+  'privacy.cookies.intro': 'The following data is stored locally in your browser:',
+  'privacy.cookies.item1': 'Language preference (localStorage) — remembers whether you chose the Dutch or English version of the site.',
+  'privacy.cookies.item2': 'Dark mode (localStorage) — remembers whether you chose the light or dark theme.',
+  'privacy.contact.title': 'Contact',
+  'privacy.contact.text': 'Do you have questions about our privacy policy or would you like to exercise your rights? Please contact us:',
+  'privacy.contact.email': 'Email',
+
+  // Terms & Conditions
+  'terms.title': 'Terms & Conditions',
+  'terms.lastUpdated': 'Last updated',
+  'terms.intro.title': 'General Provisions',
+  'terms.intro.text': 'These terms and conditions apply to membership of Shi-Sei Sport. By becoming a member you accept these conditions.',
+
+  'terms.expect.title': 'What you can expect from Shi-Sei Sport',
+  'terms.expect.item1': 'Responsible sports in a safe environment.',
+  'terms.expect.item2': 'Your personal and address details entrusted to us will not be shared with third parties without consulting you.',
+  'terms.expect.item3': 'Lessons are taught by skilled, trained and/or enthusiastic coaches.',
+  'terms.expect.item4': 'All supervisors, volunteers and/or coaches hold a valid Certificate of Good Conduct (VOG).',
+  'terms.expect.item5': 'Each department has its own dedicated professional coach/supervisor.',
+  'terms.expect.item6': 'A professionally run sports club with adequate first aid resources.',
+  'terms.expect.item7': 'A great deal of enthusiasm from every leader.',
+
+  'terms.enrollment.title': 'Enrollment',
+  'terms.enrollment.item1': 'Please read the terms and conditions carefully before completing and submitting the enrollment form.',
+  'terms.enrollment.item2pre': 'Every member is required to take note of the',
+  'terms.enrollment.regelementLink': 'general regulations',
+  'terms.enrollment.huishoudelijkLink': 'house regulations',
+  'terms.enrollment.item2mid': 'and the',
+  'terms.enrollment.gedragscodeLink': 'code of conduct',
+
+  'terms.signature.title': 'Agreement',
+  'terms.signature.item1': 'By checking or signing the online enrollment form, the undersigned declares to agree with the membership conditions of Shi-Sei Sport, of which the General Terms and the dojo rules are an inseparable part, and understands what is stated therein.',
+  'terms.signature.item2': 'The General Terms are visible on the website www.shi-sei.nl.',
+  'terms.signature.item3': 'The rights arising from the membership are strictly personal.',
+  'terms.signature.item4': 'Persons under 18 years of age may only be registered by their parent(s) or legal guardian(s).',
+  'terms.signature.item5': 'By signing the direct debit authorization, you grant Shi-Sei Sport permission to automatically debit the amounts due for advance payment of the membership fee and other amounts owed to Shi-Sei Sport from your specified bank account until further notice.',
+
+  'terms.membership.title': 'Membership',
+  'terms.membership.item1': 'Membership commences upon receipt of registration and payment of the membership fee.',
+  'terms.membership.item2': 'Persons under 18 years of age may only be registered by their parent(s) or legal guardian(s).',
+  'terms.membership.item3pre': 'Members must adhere to the',
+  'terms.membership.item3mid': 'and',
+  'terms.membership.item4': 'Membership is personal and non-transferable.',
+
+  'terms.link.huisregels': 'house rules',
+  'terms.link.gedragscode': 'code of conduct',
+
+  'terms.rates.title': 'Pricing',
+  'terms.rates.item1': 'Rates may be adjusted by Shi-Sei Sport if necessary.',
+  'terms.rates.item2': 'Changes to membership rates will be communicated by email 1 month before they take effect.',
+  'terms.rates.item3': 'In the event of a price increase, the member has the right to terminate the agreement within 4 weeks of the announcement.',
+  'terms.rates.item4': 'Any prepaid subscription fees will be refunded.',
+
+  'terms.payment.title': 'Payment',
+  'terms.payment.item1': 'The member or their representative is responsible for ensuring smooth payment of membership contributions.',
+  'terms.payment.item2': 'Membership contributions are automatically collected quarterly around the first of the new month.',
+  'terms.payment.item3': 'Contributions are paid quarterly in advance.',
+  'terms.payment.item4': 'In case of late payment, access to classes may be suspended.',
+  'terms.payment.item5': 'Refund of membership fees is not possible, except in exceptional cases.',
+
+  'terms.behavior.title': 'Behavior and House Rules',
+  'terms.behavior.item1': 'Members must treat other members, trainers and equipment with respect.',
+  'terms.behavior.item2pre': 'Members must adhere to the',
+  'terms.behavior.item2mid': 'and',
+  'terms.behavior.item3': 'Participation in training under the influence of alcohol or drugs is not permitted.',
+  'terms.behavior.item4': 'In case of serious violations, the board may decide to suspend or terminate membership.',
+
+  'terms.exams.title': 'Exams and Competitions',
+  'terms.exams.item1': 'The costs for the special competition group, competitions and exams are not included in the lesson fees.',
+  'terms.exams.item2': 'In the event of absence of a member at an exam or competition, no refund of the registration fee will be given.',
+
+  'terms.liability.title': 'Own Risk',
+  'terms.liability.text1': 'Shi-Sei Sport is not liable for injury, loss or damage that occurs during or as a result of sports activities, unless this is the result of gross negligence or intent by the association.',
+  'terms.liability.text2': 'Members are strongly advised to take out their own sports injury or accident insurance.',
+
+  'terms.photos.title': 'Photos and Media',
+  'terms.photos.item1': 'The member and other visitors agree to the use of their image in print, photo, video, etc. for the website and other promotional purposes without claiming compensation.',
+  'terms.photos.item2': 'If you object to this, you can indicate this in writing on the enrollment form and communicate any changes by email at any time.',
+
+  'terms.changes.title': 'Changes',
+  'terms.changes.text': 'Shi-Sei Sport reserves the right to change these terms and conditions. Members will be notified of this via the website or by email.',
+
+  'terms.cancellation.title': 'Cancellation',
+  'terms.cancellation.item1': 'Membership can be terminated with one month\'s notice.',
+  'terms.cancellation.item2': 'Cancellation must be done in writing or by email.',
+  'terms.cancellation.item3': 'If cancelled within the contract period, membership fees remain due until the end of that period.',
+
+  'terms.contact.title': 'Questions',
+  'terms.contact.text': 'Do you have questions about these terms and conditions? Please contact us:',
+  'terms.contact.email': 'Email',
+
+  // Common
+  'common.loading': 'Loading...',
+  'common.required': '*',
+  'common.select': 'Select',
+  'common.optional': '(optional)',
+  'common.invalidEmail': 'Please enter a valid email address, e.g. name@example.com',
+  'common.invalidPhone': 'Please enter a valid phone number, e.g. +31-612345678',
+  'common.invalidIban': 'Please enter a valid IBAN, e.g. NL91 ABNA 0417 1643 00',
+  'common.invalidPostalCode': 'Please enter a valid postal code, e.g. 1234 AB',
+  'common.invalidAge': 'Please enter a valid age (4\u2013100)',
+  'common.ibanCountryHint': 'The form automatically adjusts to the country code.',
+
+  // Placeholders
+  'placeholder.name': 'Your full name',
+  'placeholder.firstName': 'First name',
+  'placeholder.middleName': 'Prefix (optional)',
+  'placeholder.lastName': 'Last name',
+  'placeholder.email': 'name@example.com',
+  'placeholder.phone': '06-12345678',
+  'placeholder.street': 'Street name',
+  'placeholder.houseNumber': 'No.',
+  'placeholder.city': 'City',
+  'placeholder.accountHolder': 'Account holder name',
+  'placeholder.age': 'e.g. 8',
+  'placeholder.message': 'Type your message here...',
+
+  // 404
+  '404.title': 'Page not found',
+  '404.description': 'The page you are looking for does not exist or has been moved.',
+  '404.home': 'Go to Homepage',
+  '404.back': 'Back',
+};
+
+export default en;
