+++ frontend/src/i18n/nl.ts
@@ -1,0 +1,560 @@
+const nl: Record<string, string> = {
+  // Navigation
+  'nav.home': 'Home',
+  'nav.about': 'Over Ons',
+  'nav.team': 'Het Team',
+  'nav.history': 'Historie',
+  'nav.locations': 'Locaties',
+  'nav.rules': 'VCP en Regels',
+  'nav.info': 'Informatie',
+  'nav.schedule': 'Rooster',
+  'nav.pricing': 'Tarieven',
+  'nav.exam': 'Examen Eisen',
+  'nav.enrollment': 'Inschrijven',
+  'nav.agenda': 'Agenda',
+  'nav.news': 'Nieuws',
+  'nav.media': 'Media',
+  'nav.contact': 'Contact',
+  'nav.trial': 'Gratis Proefles',
+
+  // Hero
+  'hero.title': 'Judo',
+  'hero.subtitle': 'SHI-SEI SPORT',
+  'hero.description': 'De oudste judo club van Den Haag',
+  'hero.text': 'Wij geven judo les op maandag, woensdag, donderdag en zaterdag. Ervaar de kracht van discipline, respect en zelfvertrouwen in onze dojo.',
+  'hero.button': 'Probeer een Proefles',
+  'hero.scheduleButton': 'Bekijk het Rooster',
+
+  // News
+  'news.latest': 'Laatste Nieuws',
+  'news.search': 'Zoek nieuws...',
+  'news.all': 'Alle nieuws',
+  'news.readMore': 'Lees meer',
+  'news.clickToRead': 'Klik om het volledige bericht te lezen...',
+  'news.loading': 'Laden...',
+  'news.notFound': 'Nieuws niet gevonden',
+  'news.back': 'Terug naar nieuws',
+  'news.backAll': 'Terug naar alle nieuws',
+  'news.noNews': 'Er is nog geen nieuws beschikbaar.',
+  'news.title': 'Laatste Nieuws',
+  'news.description': 'Blijf op de hoogte van het laatste nieuws en evenementen bij Shi-Sei Sport.',
+  'news.share': 'Deel dit artikel',
+  'news.shareCopied': 'Link gekopieerd!',
+  'news.noContent': 'Dit artikel heeft geen inhoud.',
+
+  // Schedule
+  'schedule.title': 'Lessen & Schema',
+  'schedule.description': 'Wij geven judo les op maandag, woensdag, donderdag en zaterdag',
+  'schedule.loading': 'Rooster laden...',
+  'schedule.error': 'Kon het rooster niet laden. Probeer het later opnieuw.',
+  'schedule.retry': 'Opnieuw proberen',
+  'schedule.noClasses': 'Er zijn nog geen lessen ingepland.',
+
+  // Locations
+  'locations.title': 'Onze Trainingslocaties',
+  'locations.address': 'Adres',
+  'locations.noMap': 'Geen kaart beschikbaar',
+  'locations.noLocations': 'Geen locaties beschikbaar',
+  'locations.loading': 'Locaties laden...',
+  'locations.error': 'Kon de locaties niet laden. Probeer het later opnieuw.',
+  'locations.questionsTitle': 'Vragen over onze locaties?',
+  'locations.postalAddress': 'Postadres',
+  'locations.phone': 'Telefoon',
+  'locations.email': 'Email',
+  'locations.hours': 'Openingstijden',
+  'locations.hoursText': 'Afhankelijk van lesrooster',
+  'locations.contactTitle': 'Vragen of Contact Nodig?',
+  'locations.contactDescription': 'Neem contact met ons op voor meer informatie over onze trainingslocaties, openingstijden of andere vragen.',
+  'locations.contactButton': 'Ga naar Contactpagina',
+  'locations.openMaps': 'Open in Google Maps',
+
+  // Contact
+  'contact.title': 'Neem Contact Op',
+  'contact.description': 'Heeft u vragen? Wij staan voor u klaar! Vul het formulier in of neem direct contact met ons op.',
+  'contact.info': 'Contactgegevens',
+  'contact.address': 'Postadres',
+  'contact.phone': 'Telefoon',
+  'contact.email': 'Email',
+  'contact.hours': 'Trainingsuren',
+  'contact.hoursText': 'Maandag, Woensdag, Donderdag en Zaterdag',
+  'contact.fullScheduleText': 'Bekijk het volledige rooster op de',
+  'contact.scheduleLink': 'rooster pagina',
+  'contact.formTitle': 'Stuur een Bericht',
+  'contact.name': 'Naam',
+  'contact.subject': 'Onderwerp',
+  'contact.message': 'Bericht',
+  'contact.send': 'Verstuur Bericht',
+  'contact.sending': 'Verzenden...',
+  'contact.error': 'Er ging iets mis bij het verzenden. Probeer het opnieuw.',
+  'contact.captchaRequired': 'Voltooi de CAPTCHA verificatie voordat u het formulier verstuurt.',
+  'common.captchaLabel': 'Ik ben geen robot, maar een Judoka',
+  'contact.success': 'Bericht Ontvangen!',
+  'contact.successText': 'Bedankt voor uw bericht! Wij nemen zo spoedig mogelijk contact met u op.',
+  'contact.subjectOptions.proefles': 'Gratis Proefles',
+  'contact.subjectOptions.inschrijving': 'Inschrijving',
+  'contact.subjectOptions.vraag': 'Algemene Vraag',
+  'contact.subjectOptions.anders': 'Anders',
+  'contact.subjectSelect': 'Selecteer een onderwerp',
+
+  // Team
+  'team.title': 'Ons Team',
+  'team.description': 'Ontmoet onze ervaren en gepassioneerde trainers die zich inzetten voor uw ontwikkeling in judo.',
+  'team.cta': 'Interesse in Training?',
+  'team.ctaText': 'Kom langs voor een gratis proefles en maak kennis met ons team!',
+  'team.button': 'Boek een Gratis Proefles',
+  'team.error': 'Er is een fout opgetreden bij het laden van het team.',
+  'team.noInstructors': 'Geen instructeurs gevonden.',
+  'team.qualifications': 'Certificeringen & Prestaties:',
+
+  // History
+  'history.title': 'Onze Rijke Geschiedenis',
+  'history.section1.title': 'De Oprichting (1950)',
+  'history.section1.p1': 'In november 1950 werd JUDOVERENIGING SHI-SEI opgericht. Een aantal enthousiaste judoka\'s van Johan v.d. Bruggen uit de Zoutmanstraat wilden een eigen club. Een club waar men goedkoper en ook meer kon trainen.',
+  'history.section1.p2': 'Onder leiding van Jacques Brakel, Hans van Diggelen en nog enkele anderen, werd op 5 november 1950 de Judovereniging SHI-SEI geboren, aangemeld bij de Kamer van Koophandel en bij de Nederlandse Amateur Judo Associatie aangesloten.',
+  'history.section1.p3': 'In de beginperiode werd geoefend in een gymzaal aan de Daal en Bergselaan, maar al vrij snel verhuisde men naar een permanente dojo in de Jan Hendrikstraat 9. Hier kon naar hartelust worden geoefend en zorgde de NAJA-bondstrainer Tokio Hirano, 6e dan, voor verbetering van de technieken. Er kon nu 7 dagen in de week worden getraind en desnoods de gehele dag. Het duurde dan ook niet lang of de trainingen brachten resultaat.',
+  'history.section1.tokioHiranoLink': 'https://judoinfo.com/hirano/#:~:text=Tokio%20Hirano:%20The%20Man%20Who%20Revolutionized%20Judo%20%7C%20Judo%20Info.%20Judo%20Info.%20Online%20Dojo.',
+  'history.section1.p4': 'Wil Wagner was de eerste zwartebandhouder van de club. Op een judo-zomerschool in Zwitserland ontving hij uit handen van grootmeester Hanno Rhi de fel begeerde band. Kort daarna behaalde Jan v.d. Toorn als tweede ook de 1e dan graad (zwarte band). Zij beiden waren in de beginperiode de technische leiding op de mat, aangevuld met bruine banden die fungeerden als oefenmeester op de diverse avonden.',
+  'history.section1.p5': 'De organisatorische leiding onder aanvoering van voorzitter Jacques Brakel zorgde dat de club ook in dat opzicht uitstekend draaide. Zwarte- en bruine banden werden ingeschreven voor Oefenmeester- en scheidsrechterscursussen en zorgden op die wijze voor de nodige "know-how".',
+  'history.growth.title': 'Groei in de Jaren Zestig',
+  'history.growth.description': 'In de jaren zestig was SHI-SEI een toonaangevende club in Den Haag en een geduchte tegenstander op de mat. Steeds groter werd het ledenaantal en ook het kader breidde zich al maar uit.',
+  'history.growth.didYouKnow': 'Wist je dat SHI-SEI de moederclub werd van vele Haagse judoclubs zoals Lu-Gia-Jen, Guan-Tsui, Huan-Pi, De Doorkruiers, Guan Lin, Kagawa en Sumeju? Wil Wagner was de grondlegger van de meeste van deze clubs.',
+  'history.section2.title': 'Succesvolle Judoka\'s',
+  'history.section2.p1': 'Honderden zwartebandhouders heeft de club inmiddels afgeleverd waaronder een aantal bijzonder goede wedstrijdjudoka zoals Wil Tap, Jan Hendriks, Jaap Zoutendijk, Jan Dotinga, Kees Toonen, Koos Tap, Koos Letterie, Wim Lut, Giel Groos, Beb de Waal, Nico Schoemaker, Koos v.d. Mark, Wil Wagner, Jan v.d. Toorn, Ed Beekink, Wim de Rooy, Barry Oudshoorn, Henk Ammerlaan, John Lut en nog vele anderen.',
+  'history.section3.title': 'Wederopstanding & Nieuwe Locaties',
+  'history.section3.p1': 'In 1970 ging de vereniging bijna ter ziele omdat een aantal leden op een buitengewone ledenvergadering het voltallige bestuur aan de kant zette. Dit liep uit op een falicante mislukking.',
+  'history.section3.p2': 'Doch in 1972 werd onder de bezielende leiding van Wim Lut een nieuw bestuur gevormd en werd vooral door vestigingen in het Westland de club weer nieuw leven ingeblazen. Dit had echter tot gevolg dat de accommodatie veel te klein werd en niet meer voldeed aan de eisen van die tijd.',
+  'history.section3.p3': 'Eindelijk in 1981 kreeg men door toedoen van wethouders Piet Vink en Adrie Duivestein de beschikking over een ruime accommodatie in de oude school aan de Nieuwe Schoolstraat 22-B.',
+  'history.section3.p4': 'Hier hebben we de beschikking over een ruime oefenzaal, twee ruime douche-gelegenheden, kantoor, Kantine, kleine oefenzaal en heel veel bergruimte. In deze ruimte hebben we nu niet alleen judo maar ook al sinds 1990 een taekwondo-afdeling met diverse eigen-kweek 1e en 2e dans. Ook een dart-afdeling zorgt voor de nodige gezelligheid.',
+  'history.section4.title': 'Verhuizingen & Recente Historie',
+  'history.section4.p1': 'Helaas kwam in 2006 aan onze mooie eigen locatie een einde toen de gemeente besloot om ons huur contract op te zeggen om een drietal luxe woningen te bouwen op de plek van onze prachtige locatie. We konden helaas geen eigen nieuwe locatie vinden die wij konden bekostigen.',
+  'history.section4.p2': 'Wij besloten gebruik te maken van een buurthuis in de Fluitenbergstraat te Den Haag en een gymzaal aan de Beresteinlaan te Den Haag. In 2011 werd ontmoetingscentrum Morgenstond onze nieuwe locatie en werd de Gymzaal aan de Beresteinlaan wederom noodgedwongen ingeruild voor een mooiere gymzaal aan de Pachtersdreef.',
+  'history.section4.p3': 'Vanaf 2011 tot aan de corona crisis in 2020 gaven wij 6 dagen per week judo en taekwondo lessen. Ons ledental was gestaagd gegroeid naar ongeveer 140 judoleden en 40 taekwondoleden.',
+  'history.section4.p4': 'Vanaf 16 maart 2020 waren wij noodgedwongen te sluiten voor een korte periode i.v.m. met de corona virus infectie. Maar inmiddels zijn we weer op volle toeren hardop aan het trainen.',
+  'history.timeline.title': 'Tijdlijn',
+  'history.stats.title': 'Vandaag de dag',
+  'history.stats.members': 'Leden',
+  'history.stats.region': 'Locatie',
+  'history.stats.city': 'Den Haag',
+  'history.milestones.1950.title': 'Oprichting',
+  'history.milestones.1950.description': 'Start aan de Daal en Bergselaan.',
+  'history.milestones.1960.title': 'Gouden Jaren',
+  'history.milestones.1960.description': 'Moederclub van vele Haagse verenigingen.',
+  'history.milestones.1980.title': 'Nieuwe Schoolstraat',
+  'history.milestones.1980.description': 'Verhuizing naar de iconische dojo (1981).',
+  'history.milestones.2011.title': 'Morgenstond',
+  'history.milestones.2011.description': 'Nieuwe locatie aan de Pachtersdreef.',
+  'history.milestones.2025.title': 'Heden',
+  'history.milestones.2025.description': 'Nog steeds actief in Den Haag met 150+ leden.',
+
+  // Pricing
+  'pricing.title': 'Lidmaatschap & Tarieven',
+  'pricing.popular': 'Populair',
+  'pricing.month': 'per maand',
+  'pricing.plans.youth.name': 'Jeugd (t/m 18 jaar)',
+  'pricing.plans.youth.price': '€27,50',
+  'pricing.plans.youth.f1': 'Onbeperkt trainen',
+  'pricing.plans.youth.f2': 'Toegang tot alle jeugdlessen',
+  'pricing.plans.youth.f3': 'Examen mogelijkheden',
+  'pricing.plans.youth.f4': 'Gratis judo pak t/m 17 jaar',
+  'pricing.plans.adults.name': 'Volwassenen',
+  'pricing.plans.adults.price': '€27,50',
+  'pricing.plans.adults.f1': '2x per week trainen',
+  'pricing.plans.adults.f2': 'Toegang tot alle lessen',
+  'pricing.plans.adults.f3': 'Examen mogelijkheden',
+  'pricing.year': 'per jaar',
+  'pricing.plans.youth.yearly': '€330,-',
+  'pricing.plans.adults.yearly': '€330,-',
+  'pricing.registrationFee': 'Eenmalige Inschrijfkosten',
+  'pricing.registrationFeeDescription': 'Deze kosten betaal je alleen bij je eerste inschrijving en gelden voor je gehele lidmaatschap.',
+  'pricing.ooievaarspas': 'Met de ooievaarspas tot 100% korting op de contributie mogelijk',
+  'pricing.startNow': 'Start Nu',
+  'pricing.loading': 'Tarieven laden...',
+  'pricing.error': 'Fout bij het laden van tarieven',
+  'pricing.noPrices': 'Geen prijsplannen beschikbaar',
+  'pricing.cta.title': 'Gratis Proefles',
+  'pricing.cta.desc': 'Probeer eerst een gratis proefles voordat u zich inschrijft. Geen verplichtingen!',
+  'pricing.cta.button': 'Boek een Gratis Proefles',
+
+  // Exam
+  'exam.title': 'Examen Eisen',
+  'exam.participationTitle': 'Deelname is altijd gewenst',
+  'exam.participationText': 'Hoewel er voor volledige bandgraden specifieke minimale eisen gelden, zijn er tijdens de examens ook altijd tussentijdse slippen te behalen. Wij moedigen daarom iedereen aan om mee te doen.',
+  'exam.loading': 'Laden...',
+  'exam.error': 'Fout bij het laden van examen eisen',
+  'exam.noGrades': 'Er zijn geen examen eisen beschikbaar.',
+  'exam.download': 'Download Examen Document',
+  'exam.preview': 'Voorbeeld',
+  'exam.supplementary': 'Aanvullende Documenten',
+  'exam.minAge': 'Minimale leeftijd',
+
+  // Agenda
+  'agenda.badge': 'Jaarkalender',
+  'agenda.title': 'Jaaroverzicht Agenda',
+  'agenda.loading': 'Laden...',
+  'agenda.error': 'Fout bij het laden van de agenda',
+  'agenda.noEvents': 'Er zijn geen evenementen gepland.',
+  'agenda.vacation': 'Vakantie',
+  'agenda.holiday': 'Feestdag',
+  'agenda.exam': 'Examen',
+  'agenda.competition': 'Wedstrijd',
+  'agenda.other': 'Evenement',
+  'agenda.allDay': 'Hele dag',
+  'agenda.multiDay': 'Meerdaagse',
+  'agenda.registrationRequired': 'Aanmelden verplicht',
+  'agenda.registrationDeadline': 'Deadline',
+  'agenda.maxParticipants': 'Max deelnemers',
+  'agenda.moreInfo': 'Meer informatie',
+  'agenda.location': 'Locatie',
+
+  // Media/Albums
+  'media.title': 'Albums',
+  'media.description': 'Bekijk foto\'s en videos van trainingen, evenementen, wedstrijden en bijzondere momenten.',
+  'media.search': 'Zoek media albums...',
+  'media.loading': 'Albums laden...',
+  'media.error': 'Fout bij het laden van de albums',
+  'media.noAlbums': 'Er zijn nog geen albums beschikbaar.',
+  'media.viewAlbum': 'Bekijk Album',
+  'media.close': 'Sluiten',
+  'media.previous': 'Vorige',
+  'media.next': 'Volgende',
+  'media.filterContent': 'Alle Albums',
+  'media.filterPhotos': 'Alleen Foto\'s',
+  'media.filterVideos': 'Met Video\'s',
+  'media.download': 'Downloaden',
+  'media.downloadAll': 'Alles downloaden',
+  'media.downloadingAll': 'Downloaden...',
+
+  // News month filter
+  'news.filterMonth': 'Alle maanden',
+  'news.month.1': 'Januari',
+  'news.month.2': 'Februari',
+  'news.month.3': 'Maart',
+  'news.month.4': 'April',
+  'news.month.5': 'Mei',
+  'news.month.6': 'Juni',
+  'news.month.7': 'Juli',
+  'news.month.8': 'Augustus',
+  'news.month.9': 'September',
+  'news.month.10': 'Oktober',
+  'news.month.11': 'November',
+  'news.month.12': 'December',
+
+  // Regels
+  'regels.title': 'VCP en Reglementen ',
+  'regels.description': 'Hier vindt u alle reglementen en huishoudelijke regels van Shi-Sei Sport. Lees deze documenten zorgvuldig door om op de hoogte te blijven van de regels en voorschriften.',
+  'regels.infoTitle': 'Belangrijke Informatie',
+  'regels.infoText': 'Alle leden en ouders/verzorgers worden geacht op de hoogte te zijn van deze reglementen. Klik op een document om meer details te lezen en gebruik de downloadknop om het PDF-bestand te downloaden.',
+  'regels.download': 'Download PDF',
+  'regels.noDocuments': 'Er zijn momenteel geen documenten beschikbaar.',
+  'regels.error': 'Kon de documenten niet laden. Probeer het later opnieuw.',
+
+  // VCP
+  'vcp.title': 'Vertrouwenscontactpersoon (VCP)',
+  'vcp.subtitle': 'Veilig en fijn sporten',
+  'vcp.contact': 'Contact',
+  'vcp.vcpName': 'VCP Naam',
+  'vcp.email': 'Email',
+  'vcp.whatDoesVcpDo': 'Wat doet de VCP?',
+  'vcp.forWhom': 'Voor wie is de VCP er?',
+  'vcp.whyContact': 'Waarom contact opnemen?',
+  'vcp.aboutVcp': 'Over de VCP',
+  'vcp.preventivePolicy': 'Preventief Beleid',
+  'vcp.crossingBehavior': 'Wat is grensoverschrijdend gedrag?',
+  'vcp.vcpTasks': 'Taken Vertrouwenscontactpersoon',
+
+  // Inschrijven
+  'inschrijven.title': 'Inschrijven',
+  'inschrijven.description': 'Word lid van Shi-Sei Sport. Kies hieronder hoe u zich wilt inschrijven.',
+  'inschrijven.downloadOption': 'PDF Downloaden',
+  'inschrijven.downloadOptionDesc': 'Download het inschrijfformulier als PDF, vul het in en lever het in bij de club.',
+  'inschrijven.onlineOption': 'Online Inschrijven',
+  'inschrijven.onlineOptionDesc': 'Vul het formulier direct online in en ontvang automatisch een ingevuld PDF document.',
+  'inschrijven.formTitle': 'Online Inschrijfformulier',
+  'inschrijven.formDescription': 'Vul onderstaand formulier in en ontvang automatisch een ingevuld PDF document.',
+  'inschrijven.infoTitle': 'Hoe werkt het?',
+  'inschrijven.step1': 'Download het formulier hieronder',
+  'inschrijven.step2': 'Vul het volledig in',
+  'inschrijven.step3': 'Lever het in bij de club of stuur het digitaal op',
+  'inschrijven.download': 'Download',
+  'inschrijven.noDocuments': 'Er zijn momenteel geen inschrijfformulieren beschikbaar.',
+  'inschrijven.error': 'Kon de formulieren niet laden. Probeer het later opnieuw.',
+  'inschrijven.ctaTitle': 'Hulp nodig bij inschrijven?',
+  'inschrijven.ctaText': 'Neem contact met ons op als u vragen heeft over het inschrijfproces of hulp nodig heeft bij het invullen van de formulieren.',
+  'inschrijven.ctaButton': 'Neem Contact Op',
+  'inschrijven.heroTagline': 'Begin jouw judoreis vandaag',
+  'inschrijven.heroSub': 'Word lid van onze club en ontdek de kracht van judo',
+
+  // Enrollment Form
+  'enrollment.form.personalInfo': 'Persoonlijke Gegevens',
+  'enrollment.form.firstName': 'Roepnaam',
+  'enrollment.form.middleName': 'Tussenvoegsel',
+  'enrollment.form.lastName': 'Achternaam',
+  'enrollment.form.dateOfBirth': 'Geboortedatum',
+  'enrollment.form.email': 'E-mail',
+  'enrollment.form.phone': 'Telefoonnummer',
+  'enrollment.form.address': 'Adres',
+  'enrollment.form.street': 'Straatnaam',
+  'enrollment.form.houseNumber': 'Huisnummer',
+  'enrollment.form.postalCode': 'Postcode',
+  'enrollment.form.city': 'Plaats',
+  'enrollment.form.emergencyContact': 'Noodcontactpersoon',
+  'enrollment.form.contactName': 'Naam',
+  'enrollment.form.contactPhone': 'Telefoonnummer',
+  'enrollment.form.contactRelation': 'Relatie',
+  'enrollment.form.relationPlaceholder': 'Bijv. ouder, partner, vriend',
+  'enrollment.form.guardianName': 'Naam ouder/voogd',
+  'enrollment.form.guardianNamePlaceholder': 'Alleen invullen bij inschrijving minderjarige',
+  'enrollment.form.judoExperience': 'Judo Ervaring',
+  'enrollment.form.experience': 'Ervaringsniveau',
+  'enrollment.form.experienceBeginner': 'Beginner',
+  'enrollment.form.experienceSome': 'Enige ervaring',
+  'enrollment.form.experienceAdvanced': 'Gevorderd',
+  'enrollment.form.judoGrade': 'Judo Graad',
+  'enrollment.form.judoGrade.none': 'Geen / Onbekend',
+  'enrollment.form.judoGrade.6kyu': '6e kyu (Wit)',
+  'enrollment.form.judoGrade.5kyu': '5e kyu (Geel)',
+  'enrollment.form.judoGrade.4kyu': '4e kyu (Oranje)',
+  'enrollment.form.judoGrade.3kyu': '3e kyu (Groen)',
+  'enrollment.form.judoGrade.2kyu': '2e kyu (Blauw)',
+  'enrollment.form.judoGrade.1kyu': '1e kyu (Bruin)',
+  'enrollment.form.judoGrade.dan': 'Dan (Zwarte band)',
+  'enrollment.form.preferredDays': 'Voorkeur Trainingsdagen',
+  'enrollment.form.day.monday': 'Maandag',
+  'enrollment.form.day.wednesday': 'Woensdag',
+  'enrollment.form.day.thursday': 'Donderdag',
+  'enrollment.form.day.saturday': 'Zaterdag',
+  'enrollment.form.medicalInfo': 'Medische Informatie',
+  'enrollment.form.medicalInfoPlaceholder': 'Allergieën, medicatie, beperkingen...',
+  'enrollment.form.remarks': 'Opmerkingen',
+  'enrollment.form.remarksPlaceholder': 'Eventuele vragen of opmerkingen...',
+  'enrollment.form.paymentMethod': 'Betaalmethode',
+  'enrollment.form.paymentMethodNote': 'Selecteer uw betaalmethode. Met de Ooievaarspas ontvangt u korting en hoeft u geen machtiging in te vullen.',
+  'enrollment.form.ooievaarspas': 'Ik heb een Ooievaarspas',
+  'enrollment.form.ooievaarspasNote': 'Vink aan als u een geldige Ooievaarspas heeft.',
+  'enrollment.form.ooievaarspasNumber': 'Ooievaarspas nummer',
+  'enrollment.form.ooievaarspasNumberPlaceholder': 'Vul uw Ooievaarspas nummer in',
+  'enrollment.form.ooievaarspasCheckLocation': 'Uw Ooievaarspas wordt op locatie gecontroleerd.',
+  'enrollment.form.machtiging': 'Machtiging (Automatische Incasso)',
+  'enrollment.form.machtigingNote': 'Voor automatische incasso hebben we uw bankgegevens nodig.',
+  'enrollment.form.accountHolder': 'Naam Rekeninghouder',
+  'enrollment.form.iban': 'IBAN Rekeningnummer',
+  'enrollment.form.submit': 'Inschrijving Versturen',
+  'enrollment.form.submitting': 'Bezig met versturen...',
+  'enrollment.form.success': 'Inschrijving Ontvangen!',
+  'enrollment.form.successText': 'Bedankt voor uw inschrijving! We hebben uw gegevens ontvangen en nemen binnen 3 werkdagen contact met u op.',
+  'enrollment.form.pdfGenerated': 'Een ingevuld PDF formulier is automatisch gegenereerd en opgeslagen.',
+  'enrollment.form.error': 'Er ging iets mis bij het versturen. Probeer het opnieuw.',
+  'enrollment.form.captchaRequired': 'Voltooi de CAPTCHA verificatie voordat u het formulier verstuurt.',
+  'enrollment.form.agreeTermsPre': 'Ik ga akkoord met de',
+  'enrollment.form.agreeTermsLink': 'Algemene Voorwaarden',
+  'enrollment.form.agreeTermsMid': 'en het',
+  'enrollment.form.agreePrivacyLink': 'Privacy Beleid',
+  'enrollment.form.agreeTermsPost': 'van Shi-Sei Sport.',
+  'enrollment.form.signature': 'Handtekening',
+  'enrollment.form.signatureClear': 'Wissen',
+  'enrollment.form.disclaimer': 'Door dit formulier in te dienen ga je akkoord met ons privacy beleid.',
+
+  // Enrollment Confirmation
+  'enrollment.confirm.title': 'Gegevens Bevestigen',
+  'enrollment.confirm.description': 'Controleer uw gegevens voordat u de inschrijving verstuurt.',
+  'enrollment.confirm.cancel': 'Terug',
+  'enrollment.confirm.submit': 'Bevestigen & Versturen',
+
+  // Trial
+  'trial.title': 'Boek een Gratis Proefles',
+  'trial.description': 'Probeer judo bij Shi-Sei Sport. Geen verplichtingen, gewoon komen en ervaren!',
+  'trial.benefits': 'Wat krijg je bij een proefles?',
+  'trial.benefit1': 'Gratis en zonder verplichtingen',
+  'trial.benefit2': 'Kennis maken met onze trainers',
+  'trial.benefit3': 'Ervaren van onze trainingsstijl',
+  'trial.benefit4': 'Informatie over lidmaatschap',
+  'trial.benefit5': 'Direct kunnen starten met trainen',
+  'trial.days': 'Trainingsdagen',
+  'trial.daysText': 'Maandag, Woensdag, Donderdag en Zaterdag',
+  'trial.scheduleLink': 'Bekijk volledige rooster',
+  'trial.bring': 'Wat mee te nemen?',
+  'trial.bring1': 'Comfortabele sportkleding',
+  'trial.bring2': 'Waterfles',
+  'trial.bring3': 'Slippers',
+  'trial.bring4': 'Judo-gi wordt de eerste keer geleend',
+  'trial.forWho': 'Voor wie?',
+  'trial.forWhoText': 'Iedereen vanaf 4 jaar is welkom! Of je nu beginner bent of al ervaring hebt, we hebben lessen voor alle niveaus.',
+  'trial.formTitle': 'Boek je Proefles',
+  'trial.age': 'Leeftijd',
+  'trial.experience': 'Ervaring',
+  'trial.experienceOptions.beginner': 'Beginner',
+  'trial.experienceOptions.some': 'Enige ervaring',
+  'trial.experienceOptions.advanced': 'Gevorderd',
+  'trial.preferredDay': 'Voorkeur trainingsdag',
+  'trial.preferredDayOptions.maandag': 'Maandag',
+  'trial.preferredDayOptions.woensdag': 'Woensdag',
+  'trial.preferredDayOptions.donderdag': 'Donderdag',
+  'trial.preferredDayOptions.zaterdag': 'Zaterdag',
+  'trial.preferredDayOptions.none': 'Geen voorkeur',
+  'trial.preferredDayAdultHint': 'Training voor volwassenen (18+) is alleen op maandag en donderdag.',
+  'trial.message': 'Opmerkingen (optioneel)',
+  'trial.messagePlaceholder': 'Heb je vragen of specifieke wensen?',
+  'trial.button': 'Boek Gratis Proefles',
+  'trial.submitting': 'Verzenden...',
+  'trial.error': 'Er is een fout opgetreden bij het verzenden. Probeer het later opnieuw.',
+  'trial.captchaRequired': 'Vul alstublieft de CAPTCHA in om door te gaan.',
+  'trial.success': 'Bedankt voor je aanmelding!',
+  'trial.successText': 'We hebben je aanvraag ontvangen en nemen binnen 3 dagen contact met je op om een geschikt tijdstip te plannen.',
+  'trial.disclaimer': 'Door te boeken ga je akkoord met onze privacy policy. We nemen binnen 3 dagen contact met je op.',
+
+  // Footer
+  'footer.club': 'Shi-Sei Sport',
+  'footer.description': 'De oudste judo club van Den Haag. Judo lessen sinds 1950.',
+  'footer.quickLinks': 'Snelle Links',
+  'footer.contact': 'Contact',
+  'footer.contactText': 'Voor vragen of meer informatie, neem contact met ons op via de contactpagina.',
+  'footer.jbn': 'Officieel aangesloten bij de JBN.',
+  'footer.partners': 'Partners',
+  'footer.ooievaarspas': 'Wij accepteren de Ooievaarspas.',
+  'footer.copyright': 'Alle rechten voorbehouden.',
+  'footer.followUs': 'Volg ons op',
+  'footer.privacy': 'Privacy Policy',
+  'footer.rules': 'Huishoudelijke Regels',
+  'footer.terms': 'Algemene Voorwaarden',
+
+  // Privacy Policy
+  'privacy.title': 'Privacy Policy',
+  'privacy.lastUpdated': 'Laatst bijgewerkt',
+  'privacy.intro.title': 'Inleiding',
+  'privacy.intro.text1': 'Shi-Sei Sport respecteert de privacy van alle bezoekers van onze website en leden van onze judoclub. In deze privacy policy leggen we uit hoe we omgaan met uw persoonsgegevens.',
+  'privacy.intro.text2': 'Deze privacy policy is van toepassing op het gebruik van de website en de diensten die worden aangeboden door Shi-Sei Sport.',
+  'privacy.collection.title': 'Welke gegevens verzamelen wij?',
+  'privacy.collection.intro': 'Wij verzamelen verschillende soorten gegevens voor verschillende doeleinden:',
+  'privacy.collection.item1': 'Contactgegevens zoals naam, e-mailadres en telefoonnummer',
+  'privacy.collection.item2': 'Leeftijd en geboortedatum (voor juiste indeling in trainingsgroepen)',
+  'privacy.collection.item3': 'Ervaring en bandniveau (voor juiste trainingsgroepen)',
+  'privacy.collection.item4': 'Communicatie via onze contactformulieren en e-mail',
+  'privacy.collection.item5': 'Foto- en beeldmateriaal (voor promotionele doeleinden zoals de website en sociale media)',
+  'privacy.usage.title': 'Hoe gebruiken wij uw gegevens?',
+  'privacy.usage.intro': 'Wij gebruiken uw persoonsgegevens voor de volgende doeleinden:',
+  'privacy.usage.item1': 'Het verwerken van uw aanmelding voor lidmaatschap of proefles',
+  'privacy.usage.item2': 'Het versturen van informatie over trainingen, evenementen en examens',
+  'privacy.usage.item3': 'Het beantwoorden van vragen en verzoeken',
+  'privacy.usage.item4': 'Het verbeteren van onze diensten en website',
+  'privacy.protection.title': 'Hoe beschermen wij uw gegevens?',
+  'privacy.protection.text1': 'Wij nemen de bescherming van uw gegevens serieus en nemen passende beveiligingsmaatregelen om misbruik, verlies en ongeautoriseerde toegang te voorkomen.',
+  'privacy.protection.text2': 'Uw gegevens worden opgeslagen op beveiligde servers en zijn alleen toegankelijk voor geautoriseerd personeel van Shi-Sei Sport.',
+  'privacy.rights.title': 'Uw rechten',
+  'privacy.rights.intro': 'U heeft de volgende rechten met betrekking tot uw persoonsgegevens:',
+  'privacy.rights.item1': 'Recht op inzage: u kunt opvragen welke gegevens wij van u verwerken',
+  'privacy.rights.item2': 'Recht op correctie: u kunt vragen om uw gegevens te corrigeren als deze onjuist zijn',
+  'privacy.rights.item3': 'Recht op verwijdering: u kunt vragen om uw gegevens te verwijderen',
+  'privacy.rights.item4': 'Recht op bezwaar: u kunt bezwaar maken tegen de verwerking van uw gegevens',
+  'privacy.cookies.title': 'Cookies',
+  'privacy.cookies.text1': 'Onze website gebruikt alleen functionele opslag die noodzakelijk is voor het goed functioneren van de website. We gebruiken geen tracking cookies of analytics die uw surfgedrag volgen.',
+  'privacy.cookies.intro': 'De volgende gegevens worden lokaal opgeslagen in uw browser:',
+  'privacy.cookies.item1': 'Taalvoorkeur (localStorage) — onthoudt of u de Nederlandse of Engelse versie van de site heeft gekozen.',
+  'privacy.cookies.item2': 'Donkere modus (localStorage) — onthoudt of u het lichte of donkere thema heeft gekozen.',
+  'privacy.contact.title': 'Contact',
+  'privacy.contact.text': 'Heeft u vragen over onze privacy policy of wilt u gebruik maken van uw rechten? Neem dan contact met ons op:',
+  'privacy.contact.email': 'E-mail',
+
+  // Terms & Conditions
+  'terms.title': 'Algemene Voorwaarden',
+  'terms.lastUpdated': 'Laatst bijgewerkt',
+  'terms.intro.title': 'Algemene Bepalingen',
+  'terms.intro.text': 'Deze algemene voorwaarden zijn van toepassing op het lidmaatschap van Shi-Sei Sport. Door lid te worden accepteert u deze voorwaarden.',
+  'terms.expect.title': 'Wat kan je van Shi-Sei Sport verwachten',
+  'terms.expect.item1': 'Verantwoord sporten in een veilige sportomgeving.',
+  'terms.expect.item2': 'Uw privé- en adresgegevens worden niet zonder overleg met u gedeeld met derden.',
+  'terms.expect.item3': 'De lessen worden gegeven door vakbekwame, opgeleide en/of enthousiaste trainers.',
+  'terms.expect.item4': 'Begeleiders, vrijwilligers en/of trainers zijn allen in het bezit van een geldige VOG-verklaring omtrent gedrag.',
+  'terms.expect.item5': 'Iedere afdeling heeft zijn eigen en vaste professionele trainer/begeleider.',
+  'terms.expect.item6': 'Professioneel werkende sportclub met voldoende aanwezige EHBO-middelen.',
+  'terms.expect.item7': 'Een grote dosis enthousiasme van iedere leidinggevende.',
+  'terms.enrollment.title': 'Inschrijven',
+  'terms.enrollment.item1': 'Lees voor het invullen en inleveren van het inschrijfformulier goed de algemene voorwaarden.',
+  'terms.enrollment.item2pre': 'Ieder lid is verplicht kennis te nemen van het',
+  'terms.enrollment.regelementLink': 'algemeen reglement',
+  'terms.enrollment.huishoudelijkLink': 'huishoudelijk reglement',
+  'terms.enrollment.item2mid': 'en de',
+  'terms.enrollment.gedragscodeLink': 'gedragscode',
+  'terms.signature.title': 'Ondertekening',
+  'terms.signature.item1': 'Door het aanvinken of het ondertekenen van het online inschrijfformulier verklaart ondergetekende zich akkoord met de lidmaatschapsvoorwaarden van Shi-Sei Sport, waarvan de Algemene voorwaarden en de dojo-regels een onverbrekelijk onderdeel zijn, en begrijpt wat hierin vermeld wordt.',
+  'terms.signature.item2': 'De Algemene voorwaarden zijn zichtbaar op de website www.shi-sei.nl.',
+  'terms.signature.item3': 'De rechten uit hoofde van het lidmaatschap zijn strikt persoonlijk.',
+  'terms.signature.item4': 'Jongeren onder de 18 jaar kunnen uitsluitend worden ingeschreven door de ouder(s) of wettelijke vertegenwoordigers.',
+  'terms.signature.item5': 'Door ondertekening van het machtigingsformulier verleen je toestemming aan Shi-Sei Sport om tot wederopzegging automatisch de kosten van de door u opgegeven bankrekening af te schrijven, de bedragen welke verschuldigd zijn voor het vooruitbetalen van de lidmaatschapsbijdrage en andere gelden die Shi-Sei Sport toekomen.',
+  'terms.membership.title': 'Lidmaatschap',
+  'terms.membership.item1': 'Het lidmaatschap gaat in na ontvangst van de inschrijving en betaling van de contributie.',
+  'terms.membership.item2': 'Jongeren onder de 18 jaar kunnen uitsluitend worden ingeschreven door de ouder(s) of wettelijke vertegenwoordigers.',
+  'terms.membership.item3pre': 'Leden dienen zich te houden aan de',
+  'terms.membership.item3mid': 'en',
+  'terms.membership.item4': 'Het lidmaatschap is persoonlijk en niet overdraagbaar.',
+  'terms.link.huisregels': 'huisregels',
+  'terms.link.gedragscode': 'gedragscode',
+  'terms.rates.title': 'Tarieven',
+  'terms.rates.item1': 'Indien noodzakelijk kunnen tarieven door Shi-Sei Sport worden aangepast.',
+  'terms.rates.item2': 'Wijziging van het lidmaatschapstarief wordt 1 maand voordat het ingaat per e-mail bekendgemaakt.',
+  'terms.rates.item3': 'Indien een prijsverhoging plaatsvindt, heeft het lid het recht om de overeenkomst binnen 4 weken na de bekendmaking te ontbinden.',
+  'terms.rates.item4': 'Het eventueel vooruitbetaalde abonnementsgeld wordt terugbetaald.',
+  'terms.payment.title': 'Betaling',
+  'terms.payment.item1': 'Het lid of zijn vertegenwoordiger dient zorg te dragen voor een probleemloze betaling van de lidmaatschapsbijdragen.',
+  'terms.payment.item2': 'De lidmaatschapsbijdragen worden per kwartaal automatisch geïncasseerd rond de eerste van de nieuwe maand.',
+  'terms.payment.item3': 'De contributie wordt per kwartaal vooruitbetaald.',
+  'terms.payment.item4': 'Bij niet-tijdige betaling kan de toegang tot lessen worden opgeschort.',
+  'terms.payment.item5': 'Restitutie van contributie is niet mogelijk, behalve in uitzonderlijke gevallen.',
+  'terms.behavior.title': 'Gedrag en Huisregels',
+  'terms.behavior.item1': 'Leden dienen respectvol om te gaan met andere leden, trainers en materiaal.',
+  'terms.behavior.item2pre': 'Leden dienen zich te houden aan de',
+  'terms.behavior.item2mid': 'en',
+  'terms.behavior.item3': 'Het is niet toegestaan onder invloed van alcohol of drugs deel te nemen aan trainingen.',
+  'terms.behavior.item4': 'Bij ernstige overtredingen kan het bestuur besluiten tot schorsing of beëindiging van het lidmaatschap.',
+  'terms.exams.title': 'Examens en Wedstrijden',
+  'terms.exams.item1': 'De kosten voor de speciale wedstrijdgroep, wedstrijden en examens zijn niet bij de lesgelden inbegrepen.',
+  'terms.exams.item2': 'In het geval van afwezigheid van het lid bij een examen of wedstrijd zal er geen restitutie plaatsvinden van inschrijfgeld.',
+  'terms.liability.title': 'Eigen Risico',
+  'terms.liability.text1': 'Shi-Sei Sport is niet aansprakelijk voor letsel, verlies of schade die tijdens of als gevolg van sportactiviteiten ontstaat, tenzij dit het gevolg is van grove nalatigheid of opzet van de vereniging.',
+  'terms.liability.text2': 'Leden worden dringend geadviseerd om een eigen sportblessure- of ongevallenverzekering af te sluiten.',
+  'terms.photos.title': 'Foto\'s en Beeldmateriaal',
+  'terms.photos.item1': 'Het lid en andere bezoekers gaan akkoord met het gebruik van zijn of haar beeltenis in druk, op foto, video e.d. voor website en andere promotionele doeleinden zonder daarvoor vergoeding te claimen.',
+  'terms.photos.item2': 'Bij bezwaar hiertegen kunt u dit schriftelijk op het inschrijfformulier kenbaar maken en altijd wijzigingen per e-mail doorgeven.',
+  'terms.changes.title': 'Wijzigingen',
+  'terms.changes.text': 'Shi-Sei Sport behoudt zich het recht voor deze algemene voorwaarden te wijzigen. Leden worden hiervan op de hoogte gesteld via de website of per e-mail.',
+  'terms.cancellation.title': 'Opzegging',
+  'terms.cancellation.item1': 'Het lidmaatschap kan met een opzegtermijn van één maand worden beëindigd.',
+  'terms.cancellation.item2': 'Opzegging dient schriftelijk of per e-mail te gebeuren.',
+  'terms.cancellation.item3': 'Bij opzegging binnen de contractperiode blijft men de contributie verschuldigd tot het einde van die periode.',
+  'terms.contact.title': 'Vragen',
+  'terms.contact.text': 'Heeft u vragen over deze algemene voorwaarden? Neem dan contact met ons op:',
+  'terms.contact.email': 'E-mail',
+
+  // Common
+  'common.loading': 'Laden...',
+  'common.required': '*',
+  'common.select': 'Selecteer',
+  'common.optional': '(optioneel)',
+  'common.invalidEmail': 'Voer een geldig e-mailadres in, bijv. naam@voorbeeld.nl',
+  'common.invalidPhone': 'Voer een geldig telefoonnummer in, bijv. +31-612345678',
+  'common.invalidIban': 'Voer een geldig IBAN in, bijv. NL91 ABNA 0417 1643 00',
+  'common.invalidPostalCode': 'Voer een geldige postcode in, bijv. 1234 AB',
+  'common.invalidAge': 'Voer een geldige leeftijd in (4–100)',
+  'common.ibanCountryHint': 'Het formulier past zich automatisch aan op de landcode.',
+
+  // Placeholders
+  'placeholder.name': 'Uw volledige naam',
+  'placeholder.firstName': 'Roepnaam',
+  'placeholder.middleName': 'Tussenvoegsel (optioneel)',
+  'placeholder.lastName': 'Achternaam',
+  'placeholder.email': 'naam@voorbeeld.nl',
+  'placeholder.phone': '06-12345678',
+  'placeholder.street': 'Straatnaam',
+  'placeholder.houseNumber': 'Nr.',
+  'placeholder.city': 'Woonplaats',
+  'placeholder.accountHolder': 'Naam rekeninghouder',
+  'placeholder.age': 'bijv. 8',
+  'placeholder.message': 'Typ hier uw bericht...',
+
+  // 404
+  '404.title': 'Pagina niet gevonden',
+  '404.description': 'De pagina die u zoekt bestaat niet of is verplaatst.',
+  '404.home': 'Naar Homepage',
+  '404.back': 'Terug',
+};
+
+export default nl;
